feat: split `TestModule` into multiple suites
