--- core/integration/module_functions_test.go
@@ -1,166 +1,0 @@
-package core
-
-import (
-	"context"
-	"strings"
-
-	"github.com/dagger/dagger/testctx"
-	"github.com/stretchr/testify/require"
-)
-
-func (ModuleSuite) TestDaggerCLIFunctions(ctx context.Context, t *testctx.T) {
-	c := connect(ctx, t)
-
-	ctr := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=test", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-import (
-	"context"
-	
-	"dagger/test/internal/dagger"
-)
-
-type Test struct{}
-
-// doc for FnA
-func (m *Test) FnA() *dagger.Container {
-	return nil
-}
-
-// doc for FnB
-func (m *Test) FnB() Duck {
-	return nil
-}
-
-type Duck interface {
-	DaggerObject
-	// quack that thang
-	Quack(ctx context.Context) (string, error)
-}
-
-// doc for FnC
-func (m *Test) FnC() *Obj {
-	return nil
-}
-
-// doc for Prim
-func (m *Test) Prim() string {
-	return "yo"
-}
-
-type Obj struct {
-	// doc for FieldA
-	FieldA *dagger.Container
-	// doc for FieldB
-	FieldB string
-	// doc for FieldC
-	FieldC *Obj
-	// doc for FieldD
-	FieldD *OtherObj
-}
-
-// doc for FnD
-func (m *Obj) FnD() *dagger.Container {
-	return nil
-}
-
-type OtherObj struct {
-	// doc for OtherFieldA
-	OtherFieldA *dagger.Container
-	// doc for OtherFieldB
-	OtherFieldB string
-	// doc for OtherFieldC
-	OtherFieldC *Obj
-	// doc for OtherFieldD
-	OtherFieldD *OtherObj
-}
-
-// doc for FnE
-func (m *OtherObj) FnE() *dagger.Container {
-	return nil
-}
-
-`,
-		)
-
-	t.Run("top-level", func(ctx context.Context, t *testctx.T) {
-		out, err := ctr.With(daggerFunctions()).Stdout(ctx)
-		require.NoError(t, err)
-		lines := strings.Split(out, "\n")
-		require.Contains(t, lines, "fn-a   doc for FnA")
-		require.Contains(t, lines, "fn-b   doc for FnB")
-		require.Contains(t, lines, "fn-c   doc for FnC")
-		require.Contains(t, lines, "prim   doc for Prim")
-	})
-
-	t.Run("top-level from subdir", func(ctx context.Context, t *testctx.T) {
-		// find-up should kick in
-		out, err := ctr.
-			WithWorkdir("/work/some/subdir").
-			With(daggerFunctions()).
-			Stdout(ctx)
-		require.NoError(t, err)
-		lines := strings.Split(out, "\n")
-		require.Contains(t, lines, "fn-a   doc for FnA")
-		require.Contains(t, lines, "fn-b   doc for FnB")
-		require.Contains(t, lines, "fn-c   doc for FnC")
-		require.Contains(t, lines, "prim   doc for Prim")
-	})
-
-	t.Run("return core object", func(ctx context.Context, t *testctx.T) {
-		out, err := ctr.With(daggerFunctions("fn-a")).Stdout(ctx)
-		require.NoError(t, err)
-		lines := strings.Split(out, "\n")
-		// just verify some of the container funcs are there, too many to be exhaustive
-		require.Contains(t, lines, "file                          Retrieves a file at the given path.")
-		require.Contains(t, lines, "as-tarball                    Returns a File representing the container serialized to a tarball.")
-	})
-
-	t.Run("return primitive", func(ctx context.Context, t *testctx.T) {
-		_, err := ctr.With(daggerFunctions("prim")).Stdout(ctx)
-		require.ErrorContains(t, err, `function "prim" returns type "STRING_KIND" with no further functions available`)
-	})
-
-	t.Run("alt casing", func(ctx context.Context, t *testctx.T) {
-		out, err := ctr.With(daggerFunctions("fnA")).Stdout(ctx)
-		require.NoError(t, err)
-		lines := strings.Split(out, "\n")
-		// just verify some of the container funcs are there, too many to be exhaustive
-		require.Contains(t, lines, "file                          Retrieves a file at the given path.")
-		require.Contains(t, lines, "as-tarball                    Returns a File representing the container serialized to a tarball.")
-	})
-
-	t.Run("return user interface", func(ctx context.Context, t *testctx.T) {
-		out, err := ctr.With(daggerFunctions("fn-b")).Stdout(ctx)
-		require.NoError(t, err)
-		lines := strings.Split(out, "\n")
-		require.Contains(t, lines, "quack   quack that thang")
-	})
-
-	t.Run("return user object", func(ctx context.Context, t *testctx.T) {
-		out, err := ctr.With(daggerFunctions("fn-c")).Stdout(ctx)
-		require.NoError(t, err)
-		lines := strings.Split(out, "\n")
-		// just verify some of the container funcs are there, too many to be exhaustive
-		require.Contains(t, lines, "field-a   doc for FieldA")
-		require.Contains(t, lines, "field-b   doc for FieldB")
-		require.Contains(t, lines, "field-c   doc for FieldC")
-		require.Contains(t, lines, "field-d   doc for FieldD")
-		require.Contains(t, lines, "fn-d      doc for FnD")
-	})
-
-	t.Run("return user object nested", func(ctx context.Context, t *testctx.T) {
-		out, err := ctr.With(daggerFunctions("fn-c", "field-d")).Stdout(ctx)
-		require.NoError(t, err)
-		lines := strings.Split(out, "\n")
-		// just verify some of the container funcs are there, too many to be exhaustive
-		require.Contains(t, lines, "other-field-a   doc for OtherFieldA")
-		require.Contains(t, lines, "other-field-b   doc for OtherFieldB")
-		require.Contains(t, lines, "other-field-c   doc for OtherFieldC")
-		require.Contains(t, lines, "other-field-d   doc for OtherFieldD")
-		require.Contains(t, lines, "fn-e            doc for FnE")
-	})
-}
