+++ core/integration/module_test.go
@@ -31,7 +31,6 @@ import (
 	"dagger.io/dagger/telemetry"
 	"github.com/dagger/dagger/cmd/codegen/introspection"
 	"github.com/dagger/dagger/core/modules"
-	"github.com/dagger/dagger/dagql/call"
 	"github.com/dagger/dagger/engine"
 	"github.com/dagger/dagger/engine/distconsts"
 	"github.com/dagger/dagger/internal/testutil"
@@ -42,1588 +41,6 @@ type ModuleSuite struct{}
 
 func TestModule(t *testing.T) {
 	testctx.Run(testCtx, t, ModuleSuite{}, Middleware()...)
-}
-
-func (ModuleSuite) TestGoInit(ctx context.Context, t *testctx.T) {
-	t.Run("from scratch", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=bare", "--sdk=go"))
-
-		out, err := modGen.
-			With(daggerQuery(`{bare{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"bare":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-	})
-
-	t.Run("reserved go.mod name", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=go", "--sdk=go"))
-
-		out, err := modGen.
-			With(daggerQuery(`{go{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"go":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-	})
-
-	t.Run("uses expected Go module name, camel-cases Dagger module name", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--source=.", "--name=My-Module", "--sdk=go"))
-
-		out, err := modGen.
-			With(daggerQuery(`{myModule{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"myModule":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-
-		generated, err := modGen.File("go.mod").Contents(ctx)
-		require.NoError(t, err)
-		require.Contains(t, generated, "module dagger/my-module")
-	})
-
-	t.Run("creates go.mod beneath an existing go.mod if context is beneath it", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		// don't use .git under /work so the context ends up being /work/ci
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithNewFile("/work/go.mod", "module foo/test\n").
-			WithNewFile("/work/foo.go", "package foo\n").
-			WithWorkdir("/work/ci").
-			With(daggerExec("init", "--name=beneathGoMod", "--sdk=go"))
-
-		out, err := modGen.
-			With(daggerQuery(`{beneathGoMod{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"beneathGoMod":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-
-		t.Run("names Go module after Dagger module", func(ctx context.Context, t *testctx.T) {
-			generated, err := modGen.Directory(".").File("go.mod").Contents(ctx)
-			require.NoError(t, err)
-			require.Contains(t, generated, "module dagger/beneath-go-mod")
-		})
-	})
-
-	t.Run("respects existing go.mod", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithExec([]string{"go", "mod", "init", "foo/test"}).
-			With(daggerExec("init", "--name=hasGoMod", "--merge", "--sdk=go"))
-
-		out, err := modGen.
-			With(daggerQuery(`{hasGoMod{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"hasGoMod":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-
-		t.Run("preserves module name", func(ctx context.Context, t *testctx.T) {
-			generated, err := modGen.File("go.mod").Contents(ctx)
-			require.NoError(t, err)
-			require.Contains(t, generated, "module foo/test")
-		})
-
-		t.Run("no new go.mod", func(ctx context.Context, t *testctx.T) {
-			_, err := modGen.File("dagger/go.mod").Contents(ctx)
-			require.ErrorContains(t, err, "no such file or directory")
-		})
-	})
-
-	t.Run("respects existing go.work", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithExec([]string{"go", "work", "init"}).
-			With(daggerExec("init", "--name=hasGoMod", "--sdk=go"))
-
-		out, err := modGen.
-			With(daggerQuery(`{hasGoMod{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"hasGoMod":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-
-		t.Run("go.work is edited", func(ctx context.Context, t *testctx.T) {
-			generated, err := modGen.File("go.work").Contents(ctx)
-			require.NoError(t, err)
-			require.Contains(t, generated, "use .\n")
-		})
-	})
-
-	t.Run("respects existing go.work with existing module", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithExec([]string{"go", "mod", "init", "foo/test"}).
-			WithExec([]string{"go", "work", "init"}).
-			WithExec([]string{"go", "work", "use", "."}).
-			With(daggerExec("init", "--name=hasGoMod", "--sdk=go", "--merge"))
-
-		out, err := modGen.
-			With(daggerQuery(`{hasGoMod{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"hasGoMod":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-
-		t.Run("go.work is edited", func(ctx context.Context, t *testctx.T) {
-			generated, err := modGen.File("go.work").Contents(ctx)
-			require.NoError(t, err)
-			require.Contains(t, generated, "use .\n")
-		})
-	})
-
-	t.Run("respects go.work for subdir if git dir", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := goGitBase(t, c).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithExec([]string{"go", "work", "init"}).
-			With(daggerExec("init", "--name=hasGoMod", "--sdk=go", "subdir"))
-
-		out, err := modGen.
-			WithWorkdir("./subdir").
-			With(daggerQuery(`{hasGoMod{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"hasGoMod":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-
-		t.Run("go.work is edited", func(ctx context.Context, t *testctx.T) {
-			generated, err := modGen.File("go.work").Contents(ctx)
-			require.NoError(t, err)
-			require.Contains(t, generated, "use ./subdir\n")
-		})
-	})
-
-	t.Run("ignores go.work for subdir", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithExec([]string{"go", "work", "init"}).
-			With(daggerExec("init", "--name=hasGoMod", "--sdk=go", "subdir"))
-
-		// we can't write to the go.work at the top-level so it should remain
-		// unedited, but we should still be able to execute the module as
-		// expected
-
-		out, err := modGen.
-			WithWorkdir("./subdir").
-			With(daggerQuery(`{hasGoMod{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"hasGoMod":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-
-		t.Run("go.work is unedited", func(ctx context.Context, t *testctx.T) {
-			generated, err := modGen.File("go.work").Contents(ctx)
-			require.NoError(t, err)
-			require.NotContains(t, generated, "use")
-		})
-	})
-
-	t.Run("respects parent go.mod if root points to it", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		generated := goGitBase(t, c).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithExec([]string{"go", "mod", "init", "foo/test"}).
-			WithNewFile("/work/foo.go", "package foo\n").
-			With(daggerExec("init", "--name=child", "--merge", "--sdk=go", "./child")).
-			WithWorkdir("/work/child").
-			// explicitly develop to see whether it makes a go.mod
-			With(daggerExec("develop")).
-			Directory("/work")
-
-		parentEntries, err := generated.Entries(ctx)
-		require.NoError(t, err)
-		require.Equal(t, []string{".git", "child", "foo.go", "go.mod", "go.sum"}, parentEntries)
-
-		childEntries, err := generated.Directory("child").Entries(ctx)
-		require.NoError(t, err)
-		require.NotContains(t, childEntries, "go.mod")
-
-		t.Run("preserves parent module name", func(ctx context.Context, t *testctx.T) {
-			goMod, err := generated.File("go.mod").Contents(ctx)
-			require.NoError(t, err)
-			require.Contains(t, goMod, "module foo/test")
-		})
-	})
-
-	t.Run("respects existing go.mod even if root points to parent that has go.mod", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		generated := goGitBase(t, c).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithExec([]string{"git", "init"}).
-			WithExec([]string{"go", "mod", "init", "foo/test"}).
-			WithNewFile("/work/foo.go", "package foo\n").
-			WithWorkdir("/work/child").
-			WithExec([]string{"go", "mod", "init", "my-mod"}).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--source=./child", "--merge", "--name=child", "--sdk=go", "./child")).
-			WithWorkdir("/work/child").
-			// explicitly develop to see whether it makes a go.mod
-			With(daggerExec("develop")).
-			Directory("/work")
-
-		parentEntries, err := generated.Entries(ctx)
-		require.NoError(t, err)
-		// no go.sum
-		require.Equal(t, []string{".git", "child", "foo.go", "go.mod"}, parentEntries)
-
-		childEntries, err := generated.Directory("child").Entries(ctx)
-		require.NoError(t, err)
-		require.Contains(t, childEntries, "go.mod")
-		require.Contains(t, childEntries, "go.sum")
-	})
-
-	t.Run("respects existing main.go", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithNewFile("/work/main.go", `
-					package main
-
-					type HasMainGo struct {}
-
-					func (m *HasMainGo) Hello() string { return "Hello, world!" }
-				`,
-			).
-			With(daggerExec("init", "--name=hasMainGo", "--sdk=go", "--source=."))
-
-		out, err := modGen.
-			With(daggerQuery(`{hasMainGo{hello}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"hasMainGo":{"hello":"Hello, world!"}}`, out)
-	})
-
-	t.Run("respects existing main.go even if it uses types not generated yet", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithNewFile("/work/main.go", `
-					package main
-					import (
-						"dagger/has-dagger-types/internal/dagger"
-					)
-
-					type HasDaggerTypes struct {}
-
-					func (m *HasDaggerTypes) Hello() *dagger.Container {
-						return dag.Container().
-							From("`+alpineImage+`").
-							WithExec([]string{"echo", "Hello, world!"})
-					}
-				`,
-			).
-			With(daggerExec("init", "--source=.", "--name=hasDaggerTypes", "--sdk=go"))
-
-		out, err := modGen.
-			With(daggerQuery(`{hasDaggerTypes{hello{stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"hasDaggerTypes":{"hello":{"stdout":"Hello, world!\n"}}}`, out)
-	})
-
-	t.Run("respects existing package without creating main.go", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithNewFile("/work/notmain.go", `package main
-
-type HasNotMainGo struct {}
-
-func (m *HasNotMainGo) Hello() string { return "Hello, world!" }
-`,
-			).
-			With(daggerExec("init", "--source=.", "--name=hasNotMainGo", "--sdk=go"))
-
-		out, err := modGen.
-			With(daggerQuery(`{hasNotMainGo{hello}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"hasNotMainGo":{"hello":"Hello, world!"}}`, out)
-	})
-
-	// Until the sdk/log package hits v1.0.0 we have to pin it to avoid breaking
-	// changes. (This also exposes that we're extremely reliant on semver.)
-	t.Run("pins the sdk/log version because it's unstable", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithNewFile("/work/main.go", `
-					package main
-					import (
-						"dagger/has-dagger-types/internal/dagger"
-					)
-
-					type HasDaggerTypes struct {}
-
-					func (m *HasDaggerTypes) Hello() *dagger.Container {
-						return dag.Container().
-							From("`+alpineImage+`").
-							WithExec([]string{"echo", "Hello, world!"})
-					}
-				`,
-			).
-			With(daggerExec("init", "--source=.", "--name=hasDaggerTypes", "--sdk=go"))
-
-		goMod, err := modGen.File("go.mod").Contents(ctx)
-		require.NoError(t, err)
-		require.Regexp(t, regexp.QuoteMeta(
-			`go.opentelemetry.io/otel/log => go.opentelemetry.io/otel/log v0.3.0`,
-		), goMod)
-		require.Regexp(t, regexp.QuoteMeta(
-			`go.opentelemetry.io/otel/sdk/log => go.opentelemetry.io/otel/sdk/log v0.3.0`,
-		), goMod)
-		require.Regexp(t, regexp.QuoteMeta(
-			`go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploggrpc => go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploggrpc v0.0.0-20240518090000-14441aefdf88`,
-		), goMod)
-		require.Regexp(t, regexp.QuoteMeta(
-			`go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploghttp => go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploghttp v0.3.0`,
-		), goMod)
-	})
-
-	t.Run("with source", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=bare", "--sdk=go", "--source=some/subdir"))
-
-		out, err := modGen.
-			With(daggerQuery(`{bare{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"bare":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-
-		sourceSubdirEnts, err := modGen.Directory("/work/some/subdir").Entries(ctx)
-		require.NoError(t, err)
-		require.Contains(t, sourceSubdirEnts, "main.go")
-
-		sourceRootEnts, err := modGen.Directory("/work").Entries(ctx)
-		require.NoError(t, err)
-		require.NotContains(t, sourceRootEnts, "main.go")
-	})
-
-	t.Run("multiple modules in go.work", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := goGitBase(t, c).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithExec([]string{"go", "work", "init"}).
-			With(daggerExec("init", "--sdk=go", "foo")).
-			With(daggerExec("init", "--sdk=go", "bar"))
-
-		generated, err := modGen.File("go.work").Contents(ctx)
-		require.NoError(t, err)
-		require.Contains(t, generated, "\t./foo\n")
-		require.Contains(t, generated, "\t./bar\n")
-
-		out, err := modGen.
-			WithWorkdir("./foo").
-			With(daggerQuery(`{foo{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"foo":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-
-		out, err = modGen.
-			WithWorkdir("./bar").
-			With(daggerQuery(`{bar{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"bar":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-	})
-
-	t.Run("fails if go.mod exists without merge flag", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithExec([]string{"go", "mod", "init", "foo/test"}).
-			With(daggerExec("init", "--name=hasGoMod", "--sdk=go"))
-
-		_, err := modGen.
-			With(daggerQuery(`{hasGoMod{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.ErrorContains(t, err, "existing go.mod does not")
-	})
-
-	t.Run("do not merge go.mod with parent", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithExec([]string{"go", "mod", "init", "foo/test"}).
-			With(daggerExec("init", "--name=bare", "--sdk=go", "--source=some/subdir"))
-
-		sourceSubdirEnts, err := modGen.Directory("/work/some/subdir").Entries(ctx)
-		require.NoError(t, err)
-		require.Contains(t, sourceSubdirEnts, "go.mod", "go.sum")
-	})
-}
-
-func (ModuleSuite) TestElixirInit(ctx context.Context, t *testctx.T) {
-	t.Run("from upstream", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=bare", "--sdk=github.com/dagger/dagger/sdk/elixir"))
-
-		out, err := modGen.
-			With(daggerQuery(`{bare{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"bare":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-	})
-
-	t.Run("from alias", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=bare", "--sdk=elixir"))
-
-		out, err := modGen.
-			With(daggerQuery(`{bare{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"bare":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-	})
-
-	t.Run("from alias with ref", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=bare", "--sdk=elixir@main"))
-
-		out, err := modGen.
-			With(daggerQuery(`{bare{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"bare":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-	})
-}
-
-func (ModuleSuite) TestPHPInit(ctx context.Context, t *testctx.T) {
-	t.Run("from upstream", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=bare", "--sdk=github.com/dagger/dagger/sdk/php"))
-
-		out, err := modGen.
-			With(daggerQuery(`{bare{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"bare":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-	})
-
-	t.Run("from alias", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=bare", "--sdk=php"))
-
-		out, err := modGen.
-			With(daggerQuery(`{bare{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"bare":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-	})
-
-	t.Run("from alias with ref", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=bare", "--sdk=php@main"))
-
-		out, err := modGen.
-			With(daggerQuery(`{bare{containerEcho(stringArg:"hello"){stdout}}}`)).
-			Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"bare":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-	})
-}
-
-func (ModuleSuite) TestInitLICENSE(ctx context.Context, t *testctx.T) {
-	t.Run("bootstraps Apache-2.0 LICENSE file if none found", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=licensed-to-ill", "--sdk=go"))
-
-		content, err := modGen.File("LICENSE").Contents(ctx)
-		require.NoError(t, err)
-		require.Contains(t, content, "Apache License, Version 2.0")
-	})
-
-	t.Run("do not boostrap LICENSE file if license is set empty", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=empty-license", "--sdk=go", "--license="))
-
-		files, err := modGen.Directory(".").Entries(ctx)
-		require.NoError(t, err)
-		require.NotContains(t, files, "LICENSE")
-	})
-
-	t.Run("do not bootstrap LICENSE file if no sdk is specified", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=no-license"))
-
-		files, err := modGen.Directory(".").Entries(ctx)
-		require.NoError(t, err)
-		require.NotContains(t, files, "LICENSE")
-
-		t.Run("do not bootstrap LICENSE file if no sdk is specified", func(ctx context.Context, t *testctx.T) {
-			modGen := modGen.With(daggerExec("develop", "--source=dagger"))
-
-			files, err := modGen.Directory(".").Entries(ctx)
-			require.NoError(t, err)
-			require.NotContains(t, files, "LICENSE")
-		})
-
-		t.Run("do not bootstrap LICENSE file if license is empty", func(ctx context.Context, t *testctx.T) {
-			modGen := modGen.With(daggerExec("develop", "--source=dagger", `--license=""`))
-
-			files, err := modGen.Directory(".").Entries(ctx)
-			require.NoError(t, err)
-			require.NotContains(t, files, "LICENSE")
-		})
-
-		t.Run("bootstrap a license after sdk is set on dagger develop", func(ctx context.Context, t *testctx.T) {
-			modGen := modGen.With(daggerExec("develop", "--sdk=go"))
-
-			content, err := modGen.File("LICENSE").Contents(ctx)
-			require.NoError(t, err)
-			require.Contains(t, content, "Apache License, Version 2.0")
-		})
-
-		t.Run("boostrap custom LICENSE file if sdk and license are specified", func(ctx context.Context, t *testctx.T) {
-			modGen := modGen.With(daggerExec("develop", "--sdk=go", `--license=MIT`))
-
-			content, err := modGen.File("LICENSE").Contents(ctx)
-			require.NoError(t, err)
-			require.Contains(t, content, "MIT License")
-		})
-	})
-
-	t.Run("creates LICENSE file in the directory specified by arg", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--name=licensed-to-ill", "--sdk=go", "./mymod"))
-
-		content, err := modGen.File("mymod/LICENSE").Contents(ctx)
-		require.NoError(t, err)
-		require.Contains(t, content, "Apache License, Version 2.0")
-	})
-
-	t.Run("does not bootstrap LICENSE file if it exists in the parent context", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := goGitBase(t, c).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithNewFile("/work/LICENSE", "doesnt matter").
-			WithWorkdir("/work/sub").
-			With(daggerExec("init", "--name=licensed-to-ill", "--sdk=go"))
-
-		// Check that the license file is not generated in the sub directory.
-		files, err := modGen.Directory("/work/sub").Entries(ctx)
-		require.NoError(t, err)
-		require.NotContains(t, files, "LICENSE")
-
-		// Check that the parent directory actually has a LICENSE file.
-		files, err = modGen.Directory("/work").Entries(ctx)
-		require.NoError(t, err)
-		require.Contains(t, files, "LICENSE")
-	})
-
-	t.Run("does not bootstrap LICENSE file if it exists in an arbitrary parent dir", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithNewFile("/work/LICENSE", "doesnt matter").
-			WithWorkdir("/work/sub").
-			With(daggerExec("init", "--name=licensed-to-ill", "--sdk=go"))
-
-		content, err := modGen.File("LICENSE").Contents(ctx)
-		require.NoError(t, err)
-		require.Contains(t, content, "Apache License, Version 2.0")
-	})
-
-	t.Run("bootstraps a LICENSE file when requested, even if it exists in the parent dir", func(ctx context.Context, t *testctx.T) {
-		c := connect(ctx, t)
-
-		modGen := goGitBase(t, c).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			WithNewFile("/work/LICENSE", "doesnt matter").
-			WithWorkdir("/work/sub").
-			With(daggerExec("init", "--name=licensed-to-ill", "--sdk=go", "--license=MIT"))
-
-		content, err := modGen.File("LICENSE").Contents(ctx)
-		require.NoError(t, err)
-		require.Contains(t, content, "MIT License")
-	})
-}
-
-func (ModuleSuite) TestGit(ctx context.Context, t *testctx.T) {
-	type testCase struct {
-		sdk               string
-		gitGeneratedFiles []string
-		gitIgnoredFiles   []string
-	}
-	for _, tc := range []testCase{
-		{
-			sdk: "go",
-			gitGeneratedFiles: []string{
-				"/dagger.gen.go",
-				"/internal/dagger/**",
-				"/internal/querybuilder/**",
-				"/internal/telemetry/**",
-			},
-			gitIgnoredFiles: []string{
-				"/dagger.gen.go",
-				"/internal/dagger",
-				"/internal/querybuilder",
-				"/internal/telemetry",
-			},
-		},
-		{
-			sdk: "python",
-			gitGeneratedFiles: []string{
-				"/sdk/**",
-			},
-			gitIgnoredFiles: []string{
-				"/sdk",
-			},
-		},
-		{
-			sdk: "typescript",
-			gitGeneratedFiles: []string{
-				"/sdk/**",
-			},
-			gitIgnoredFiles: []string{
-				"/sdk",
-			},
-		},
-	} {
-		tc := tc
-		t.Run(fmt.Sprintf("module %s git", tc.sdk), func(ctx context.Context, t *testctx.T) {
-			c := connect(ctx, t)
-
-			modGen := goGitBase(t, c).
-				With(daggerExec("init", "--name=bare", "--sdk="+tc.sdk))
-
-			out, err := modGen.
-				With(daggerQuery(`{bare{containerEcho(stringArg:"hello"){stdout}}}`)).
-				Stdout(ctx)
-			require.NoError(t, err)
-			require.JSONEq(t, `{"bare":{"containerEcho":{"stdout":"hello\n"}}}`, out)
-
-			t.Run("configures .gitattributes", func(ctx context.Context, t *testctx.T) {
-				ignore, err := modGen.File(".gitattributes").Contents(ctx)
-				require.NoError(t, err)
-				for _, fileName := range tc.gitGeneratedFiles {
-					require.Contains(t, ignore, fmt.Sprintf("%s linguist-generated\n", fileName))
-				}
-			})
-
-			t.Run("configures .gitignore", func(ctx context.Context, t *testctx.T) {
-				ignore, err := modGen.File(".gitignore").Contents(ctx)
-				require.NoError(t, err)
-				for _, fileName := range tc.gitIgnoredFiles {
-					require.Contains(t, ignore, fileName)
-				}
-			})
-
-			t.Run("does not configure .gitignore if disabled", func(ctx context.Context, t *testctx.T) {
-				modGen := goGitBase(t, c).
-					With(daggerExec("init", "--name=bare"))
-
-				// TODO: use dagger config to set this once support is added there
-				modCfgContents, err := modGen.File("dagger.json").Contents(ctx)
-				require.NoError(t, err)
-				modCfg := &modules.ModuleConfig{}
-				require.NoError(t, json.Unmarshal([]byte(modCfgContents), modCfg))
-				autoGitignore := false
-				modCfg.Codegen = &modules.ModuleCodegenConfig{
-					AutomaticGitignore: &autoGitignore,
-				}
-				modCfgBytes, err := json.Marshal(modCfg)
-				require.NoError(t, err)
-
-				modGen = modGen.
-					WithNewFile("dagger.json", string(modCfgBytes)).
-					With(daggerExec("develop", "--sdk=go"))
-
-				_, err = modGen.File(".gitignore").Contents(ctx)
-				require.ErrorContains(t, err, "no such file or directory")
-			})
-		})
-	}
-}
-
-//go:embed testdata/modules/go/minimal/main.go
-var goSignatures string
-
-func (ModuleSuite) TestGoSignatures(ctx context.Context, t *testctx.T) {
-	c := connect(ctx, t)
-
-	modGen := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", goSignatures)
-
-	t.Run("func Hello() string", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{hello}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"hello":"hello"}}`, out)
-	})
-
-	t.Run("func Echo(string) string", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echo(msg: "hello")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echo":"hello...hello...hello..."}}`, out)
-	})
-
-	t.Run("func EchoPointer(*string) string", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoPointer(msg: "hello")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoPointer":"hello...hello...hello..."}}`, out)
-	})
-
-	t.Run("func EchoPointerPointer(**string) string", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoPointerPointer(msg: "hello")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoPointerPointer":"hello...hello...hello..."}}`, out)
-	})
-
-	t.Run("func EchoOptional(string) string", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoOptional(msg: "hello")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptional":"hello...hello...hello..."}}`, out)
-		out, err = modGen.With(daggerQuery(`{minimal{echoOptional}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptional":"default...default...default..."}}`, out)
-	})
-
-	t.Run("func EchoOptionalPointer(string) string", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoOptionalPointer(msg: "hello")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptionalPointer":"hello...hello...hello..."}}`, out)
-		out, err = modGen.With(daggerQuery(`{minimal{echoOptionalPointer}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptionalPointer":"default...default...default..."}}`, out)
-	})
-
-	t.Run("func EchoOptionalSlice([]string) string", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoOptionalSlice(msg: ["hello", "there"])}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptionalSlice":"hello+there...hello+there...hello+there..."}}`, out)
-		out, err = modGen.With(daggerQuery(`{minimal{echoOptionalSlice}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptionalSlice":"foobar...foobar...foobar..."}}`, out)
-	})
-
-	t.Run("func Echoes([]string) []string", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoes(msgs: ["hello"])}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoes":["hello...hello...hello..."]}}`, out)
-	})
-
-	t.Run("func EchoesVariadic(...string) string", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoesVariadic(msgs: ["hello"])}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoesVariadic":"hello...hello...hello..."}}`, out)
-	})
-
-	t.Run("func HelloContext(context.Context) string", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{helloContext}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"helloContext":"hello context"}}`, out)
-	})
-
-	t.Run("func EchoContext(context.Context, string) string", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoContext(msg: "hello")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoContext":"ctx.hello...ctx.hello...ctx.hello..."}}`, out)
-	})
-
-	t.Run("func HelloStringError() (string, error)", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{helloStringError}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"helloStringError":"hello i worked"}}`, out)
-	})
-
-	t.Run("func HelloVoid()", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{helloVoid}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"helloVoid":null}}`, out)
-	})
-
-	t.Run("func HelloVoidError() error", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{helloVoidError}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"helloVoidError":null}}`, out)
-	})
-
-	t.Run("func EchoOpts(string, string, int) error", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoOpts(msg: "hi")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOpts":"hi"}}`, out)
-
-		out, err = modGen.With(daggerQuery(`{minimal{echoOpts(msg: "hi", suffix: "!", times: 2)}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOpts":"hi!hi!"}}`, out)
-	})
-
-	t.Run("func EchoOptsInline(struct{string, string, int}) error", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoOptsInline(msg: "hi")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptsInline":"hi"}}`, out)
-
-		out, err = modGen.With(daggerQuery(`{minimal{echoOptsInline(msg: "hi", suffix: "!", times: 2)}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptsInline":"hi!hi!"}}`, out)
-	})
-
-	t.Run("func EchoOptsInlinePointer(*struct{string, string, int}) error", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoOptsInlinePointer(msg: "hi")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptsInlinePointer":"hi"}}`, out)
-
-		out, err = modGen.With(daggerQuery(`{minimal{echoOptsInlinePointer(msg: "hi", suffix: "!", times: 2)}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptsInlinePointer":"hi!hi!"}}`, out)
-	})
-
-	t.Run("func EchoOptsInlineCtx(ctx, struct{string, string, int}) error", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoOptsInlineCtx(msg: "hi")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptsInlineCtx":"hi"}}`, out)
-
-		out, err = modGen.With(daggerQuery(`{minimal{echoOptsInlineCtx(msg: "hi", suffix: "!", times: 2)}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptsInlineCtx":"hi!hi!"}}`, out)
-	})
-
-	t.Run("func EchoOptsInlineTags(struct{string, string, int}) error", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoOptsInlineTags(msg: "hi")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptsInlineTags":"hi"}}`, out)
-
-		out, err = modGen.With(daggerQuery(`{minimal{echoOptsInlineTags(msg: "hi", suffix: "!", times: 2)}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptsInlineTags":"hi!hi!"}}`, out)
-	})
-
-	t.Run("func EchoOptsPragmas(string, string, int) error", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(`{minimal{echoOptsPragmas(msg: "hi")}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"echoOptsPragmas":"hi...hi...hi..."}}`, out)
-	})
-}
-
-func (ModuleSuite) TestGoSignaturesBuiltinTypes(ctx context.Context, t *testctx.T) {
-	c := connect(ctx, t)
-
-	modGen := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-import (
-	"context"
-	"dagger/minimal/internal/dagger"
-)
-
-type Minimal struct {}
-
-func (m *Minimal) Read(ctx context.Context, dir dagger.Directory) (string, error) {
-	return dir.File("foo").Contents(ctx)
-}
-
-func (m *Minimal) ReadPointer(ctx context.Context, dir *dagger.Directory) (string, error) {
-	return dir.File("foo").Contents(ctx)
-}
-
-func (m *Minimal) ReadSlice(ctx context.Context, dir []dagger.Directory) (string, error) {
-	return dir[0].File("foo").Contents(ctx)
-}
-
-func (m *Minimal) ReadVariadic(ctx context.Context, dir ...dagger.Directory) (string, error) {
-	return dir[0].File("foo").Contents(ctx)
-}
-
-func (m *Minimal) ReadOptional(
-	ctx context.Context,
-	dir *dagger.Directory, // +optional
-) (string, error) {
-	if dir != nil {
-		return dir.File("foo").Contents(ctx)
-	}
-	return "", nil
-}
-			`,
-		)
-
-	out, err := modGen.With(daggerQuery(`{directory{withNewFile(path: "foo", contents: "bar"){id}}}`)).Stdout(ctx)
-	require.NoError(t, err)
-	dirID := gjson.Get(out, "directory.withNewFile.id").String()
-
-	t.Run("func Read(ctx, Directory) (string, error)", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(fmt.Sprintf(`{minimal{read(dir: "%s")}}`, dirID))).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"read":"bar"}}`, out)
-	})
-
-	t.Run("func ReadPointer(ctx, *dagger.Directory) (string, error)", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(fmt.Sprintf(`{minimal{readPointer(dir: "%s")}}`, dirID))).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"readPointer":"bar"}}`, out)
-	})
-
-	t.Run("func ReadSlice(ctx, []Directory) (string, error)", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(fmt.Sprintf(`{minimal{readSlice(dir: ["%s"])}}`, dirID))).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"readSlice":"bar"}}`, out)
-	})
-
-	t.Run("func ReadVariadic(ctx, ...Directory) (string, error)", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(fmt.Sprintf(`{minimal{readVariadic(dir: ["%s"])}}`, dirID))).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"readVariadic":"bar"}}`, out)
-	})
-
-	t.Run("func ReadOptional(ctx, Optional[Directory]) (string, error)", func(ctx context.Context, t *testctx.T) {
-		out, err := modGen.With(daggerQuery(fmt.Sprintf(`{minimal{readOptional(dir: "%s")}}`, dirID))).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"readOptional":"bar"}}`, out)
-		out, err = modGen.With(daggerQuery(`{minimal{readOptional}}`)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, `{"minimal":{"readOptional":""}}`, out)
-	})
-}
-
-func (ModuleSuite) TestGoSignaturesUnexported(ctx context.Context, t *testctx.T) {
-	var logs safeBuffer
-	c := connect(ctx, t, dagger.WithLogOutput(&logs))
-
-	modGen := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-type Minimal struct {}
-
-type Foo struct {}
-
-type bar struct {}
-
-func (m *Minimal) Hello(name string) string {
-	return name
-}
-
-func (f *Foo) Hello(name string) string {
-	return name
-}
-
-func (b *bar) Hello(name string) string {
-	return name
-}
-`,
-		)
-
-	objs := inspectModuleObjects(ctx, t, modGen)
-	require.Equal(t, 1, len(objs.Array()))
-	require.Equal(t, "Minimal", objs.Get("0.name").String())
-
-	modGen = c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-type Minimal struct {}
-
-type Foo struct {}
-
-type bar struct {}
-
-func (m *Minimal) Hello(name string) Foo {
-	return Foo{}
-}
-
-func (f *Foo) Hello(name string) string {
-	return name
-}
-
-func (b *bar) Hello(name string) string {
-	return name
-}
-`,
-		)
-
-	objs = inspectModuleObjects(ctx, t, modGen)
-	require.Equal(t, 2, len(objs.Array()))
-	require.Equal(t, "Minimal", objs.Get("0.name").String())
-	require.Equal(t, "MinimalFoo", objs.Get("1.name").String())
-
-	modGen = c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-type Minimal struct {}
-
-type Foo struct {
-	Bar bar
-}
-
-type bar struct {}
-
-func (m *Minimal) Hello(name string) Foo {
-	return Foo{}
-}
-
-func (f *Foo) Hello(name string) string {
-	return name
-}
-
-func (b *bar) Hello(name string) string {
-	return name
-}
-`,
-		)
-
-	_, err := modGen.With(moduleIntrospection).Stderr(ctx)
-	require.Error(t, err)
-	require.NoError(t, c.Close())
-	require.Regexp(t, "cannot code-generate unexported type bar", logs.String())
-}
-
-func (ModuleSuite) TestGoSignaturesMixMatch(ctx context.Context, t *testctx.T) {
-	var logs safeBuffer
-	c := connect(ctx, t, dagger.WithLogOutput(&logs))
-
-	modGen := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-type Minimal struct {}
-
-func (m *Minimal) Hello(name string, opts struct{}, opts2 struct{}) string {
-	return name
-}
-`,
-		)
-
-	_, err := modGen.With(daggerQuery(`{minimal{hello}}`)).Stdout(ctx)
-	require.Error(t, err)
-	require.NoError(t, c.Close())
-	require.Regexp(t, "nested structs are not supported", logs.String())
-}
-
-func (ModuleSuite) TestGoSignaturesNameConflict(ctx context.Context, t *testctx.T) {
-	var logs safeBuffer
-	c := connect(ctx, t, dagger.WithLogOutput(&logs))
-
-	modGen := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-type Minimal struct {
-	Foo Foo
-	Bar Bar
-	Baz Baz
-}
-
-type Foo struct {}
-type Bar struct {}
-type Baz struct {}
-
-func (m *Foo) Hello(name string) string {
-	return name
-}
-
-func (f *Bar) Hello(name string, name2 string) string {
-	return name + name2
-}
-
-func (b *Baz) Hello() (string, error) {
-	return "", nil
-}
-`,
-		)
-
-	objs := inspectModuleObjects(ctx, t, modGen)
-	require.Equal(t, 4, len(objs.Array()))
-	require.Equal(t, "Minimal", objs.Get("0.name").String())
-	require.Equal(t, "MinimalFoo", objs.Get("1.name").String())
-	require.Equal(t, "MinimalBar", objs.Get("2.name").String())
-	require.Equal(t, "MinimalBaz", objs.Get("3.name").String())
-}
-
-func (ModuleSuite) TestGoDocs(ctx context.Context, t *testctx.T) {
-	c := connect(ctx, t)
-
-	lineEndings := map[string]string{
-		"lf":   "\n",
-		"crlf": "\r\n",
-	}
-	for name, lineEnding := range lineEndings {
-		t.Run(name, func(ctx context.Context, t *testctx.T) {
-			modGen := c.Container().From(golangImage).
-				WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-				WithWorkdir("/work").
-				With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-				WithNewFile("main.go", strings.ReplaceAll(goSignatures, "\n", lineEnding))
-
-			logGen(ctx, t, modGen.Directory("."))
-
-			obj := inspectModuleObjects(ctx, t, modGen).Get("0")
-			require.Equal(t, "Minimal", obj.Get("name").String())
-
-			hello := obj.Get(`functions.#(name="hello")`)
-			require.Equal(t, "hello", hello.Get("name").String())
-			require.Empty(t, hello.Get("description").String())
-			require.Empty(t, hello.Get("args").Array())
-
-			// test the args-based form
-			echoOpts := obj.Get(`functions.#(name="echoOpts")`)
-			require.Equal(t, "echoOpts", echoOpts.Get("name").String())
-			require.Equal(t, "EchoOpts does some opts things", echoOpts.Get("description").String())
-			require.Len(t, echoOpts.Get("args").Array(), 3)
-			require.Equal(t, "msg", echoOpts.Get("args.0.name").String())
-			require.Equal(t, "the message to echo", echoOpts.Get("args.0.description").String())
-			require.Equal(t, "suffix", echoOpts.Get("args.1.name").String())
-			require.Equal(t, "String to append to the echoed message", echoOpts.Get("args.1.description").String())
-			require.Equal(t, "times", echoOpts.Get("args.2.name").String())
-			require.Equal(t, "Number of times to repeat the message", echoOpts.Get("args.2.description").String())
-
-			// test the inline struct form
-			echoOpts = obj.Get(`functions.#(name="echoOptsInline")`)
-			require.Equal(t, "echoOptsInline", echoOpts.Get("name").String())
-			require.Equal(t, "EchoOptsInline does some opts things", echoOpts.Get("description").String())
-			require.Len(t, echoOpts.Get("args").Array(), 3)
-			require.Equal(t, "msg", echoOpts.Get("args.0.name").String())
-			require.Equal(t, "the message to echo", echoOpts.Get("args.0.description").String())
-			require.Equal(t, "suffix", echoOpts.Get("args.1.name").String())
-			require.Equal(t, "String to append to the echoed message", echoOpts.Get("args.1.description").String())
-			require.Equal(t, "times", echoOpts.Get("args.2.name").String())
-			require.Equal(t, "Number of times to repeat the message", echoOpts.Get("args.2.description").String())
-
-			// test the arg-based form (with pragmas)
-			echoOpts = obj.Get(`functions.#(name="echoOptsPragmas")`)
-			require.Equal(t, "echoOptsPragmas", echoOpts.Get("name").String())
-			require.Len(t, echoOpts.Get("args").Array(), 3)
-			require.Equal(t, "msg", echoOpts.Get("args.0.name").String())
-			require.Equal(t, "", echoOpts.Get("args.0.defaultValue").String())
-			require.Equal(t, "suffix", echoOpts.Get("args.1.name").String())
-			require.Equal(t, "String to append to the echoed message", echoOpts.Get("args.1.description").String())
-			require.Equal(t, "\"...\"", echoOpts.Get("args.1.defaultValue").String())
-			require.Equal(t, "times", echoOpts.Get("args.2.name").String())
-			require.Equal(t, "3", echoOpts.Get("args.2.defaultValue").String())
-			require.Equal(t, "Number of times to repeat the message", echoOpts.Get("args.2.description").String())
-		})
-	}
-}
-
-func (ModuleSuite) TestGoDocsEdgeCases(ctx context.Context, t *testctx.T) {
-	src := `package main
-
-// Minimal is a thing
-type Minimal struct {
-	// X is this
-	X, Y string  // Y is not this
-
-	// +private
-	Z string
-}
-
-// some docs
-func (m *Minimal) Hello(foo string, bar string,
-// hello
-baz string, qux string, x string, // lol
-) string {
-	return foo + bar
-}
-
-func (m *Minimal) HelloMore(
-	// foo here
-	foo,
-	// bar here
-	bar string,
-) string {
-	return foo + bar
-}
-
-func (m *Minimal) HelloMoreInline(opts struct{
-	// foo here
-	foo, bar string
-}) string {
-	return opts.foo + opts.bar
-}
-
-func (m *Minimal) HelloAgain( // docs for helloagain
-	foo string,
-	bar string, // docs for bar
-	baz string,
-) string {
-	return foo + bar
-}
-
-func (m *Minimal) HelloFinal(
-	foo string) string { // woops
-	return foo
-}
-
-func (m *Minimal) HelloMultiple(
-	foo string, /* foo 1 */ /* foo 2 */
-	/* bar 1 */ /* bar 2 */
-	bar string,
-) string {
-	return foo + bar
-}
-
-func (m *Minimal) HelloWeird(
-	/* foo
-data */ foo string,
-
-	bar /* bar data */ string,
-
-	/* baz data */ baz string,
-) string {
-	return foo + bar + baz
-}
-`
-
-	c := connect(ctx, t)
-
-	lineEndings := map[string]string{
-		"lf":   "\n",
-		"crlf": "\r\n",
-	}
-	for name, lineEnding := range lineEndings {
-		t.Run(name, func(ctx context.Context, t *testctx.T) {
-			modGen := c.Container().From(golangImage).
-				WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-				WithWorkdir("/work").
-				With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-				WithNewFile("main.go", strings.ReplaceAll(src, "\n", lineEnding))
-
-			logGen(ctx, t, modGen.Directory("."))
-
-			obj := inspectModuleObjects(ctx, t, modGen).Get("0")
-			require.Equal(t, "Minimal", obj.Get("name").String())
-			require.Equal(t, "Minimal is a thing", obj.Get("description").String())
-
-			hello := obj.Get(`functions.#(name="hello")`)
-			require.Equal(t, "hello", hello.Get("name").String())
-			require.Len(t, hello.Get("args").Array(), 5)
-			require.Equal(t, "foo", hello.Get("args.0.name").String())
-			require.Equal(t, "", hello.Get("args.0.description").String())
-			require.Equal(t, "bar", hello.Get("args.1.name").String())
-			require.Equal(t, "", hello.Get("args.1.description").String())
-			require.Equal(t, "baz", hello.Get("args.2.name").String())
-			require.Equal(t, "hello", hello.Get("args.2.description").String())
-			require.Equal(t, "qux", hello.Get("args.3.name").String())
-			require.Equal(t, "", hello.Get("args.3.description").String())
-			require.Equal(t, "x", hello.Get("args.4.name").String())
-			require.Equal(t, "lol", hello.Get("args.4.description").String())
-
-			hello = obj.Get(`functions.#(name="helloMore")`)
-			require.Equal(t, "helloMore", hello.Get("name").String())
-			require.Len(t, hello.Get("args").Array(), 2)
-			require.Equal(t, "foo", hello.Get("args.0.name").String())
-			require.Equal(t, "foo here", hello.Get("args.0.description").String())
-			require.Equal(t, "bar", hello.Get("args.1.name").String())
-			require.Equal(t, "bar here", hello.Get("args.1.description").String())
-
-			hello = obj.Get(`functions.#(name="helloMoreInline")`)
-			require.Equal(t, "helloMoreInline", hello.Get("name").String())
-			require.Len(t, hello.Get("args").Array(), 2)
-			require.Equal(t, "foo", hello.Get("args.0.name").String())
-			require.Equal(t, "foo here", hello.Get("args.0.description").String())
-			require.Equal(t, "bar", hello.Get("args.1.name").String())
-			require.Equal(t, "", hello.Get("args.1.description").String())
-
-			hello = obj.Get(`functions.#(name="helloAgain")`)
-			require.Equal(t, "helloAgain", hello.Get("name").String())
-			require.Len(t, hello.Get("args").Array(), 3)
-			require.Equal(t, "foo", hello.Get("args.0.name").String())
-			require.Equal(t, "", hello.Get("args.0.description").String())
-			require.Equal(t, "bar", hello.Get("args.1.name").String())
-			require.Equal(t, "docs for bar", hello.Get("args.1.description").String())
-			require.Equal(t, "baz", hello.Get("args.2.name").String())
-			require.Equal(t, "", hello.Get("args.2.description").String())
-
-			hello = obj.Get(`functions.#(name="helloFinal")`)
-			require.Equal(t, "helloFinal", hello.Get("name").String())
-			require.Len(t, hello.Get("args").Array(), 1)
-			require.Equal(t, "foo", hello.Get("args.0.name").String())
-			require.Equal(t, "", hello.Get("args.0.description").String())
-
-			hello = obj.Get(`functions.#(name="helloMultiple")`)
-			require.Equal(t, "helloMultiple", hello.Get("name").String())
-			require.Len(t, hello.Get("args").Array(), 2)
-			require.Equal(t, "foo", hello.Get("args.0.name").String())
-			require.Equal(t, "foo 1\n foo 2", hello.Get("args.0.description").String())
-			require.Equal(t, "bar", hello.Get("args.1.name").String())
-			require.Equal(t, "bar 1\n bar 2", hello.Get("args.1.description").String())
-
-			hello = obj.Get(`functions.#(name="helloWeird")`)
-			require.Equal(t, "helloWeird", hello.Get("name").String())
-			require.Len(t, hello.Get("args").Array(), 3)
-			require.Equal(t, "foo", hello.Get("args.0.name").String())
-			require.Equal(t, "foo\ndata", hello.Get("args.0.description").String())
-			require.Equal(t, "bar", hello.Get("args.1.name").String())
-			require.Equal(t, "bar data", hello.Get("args.1.description").String())
-			require.Equal(t, "baz", hello.Get("args.2.name").String())
-			require.Equal(t, "baz data", hello.Get("args.2.description").String())
-
-			require.Len(t, obj.Get(`fields`).Array(), 2)
-			prop := obj.Get(`fields.#(name="x")`)
-			require.Equal(t, "x", prop.Get("name").String())
-			require.Equal(t, "X is this", prop.Get("description").String())
-			prop = obj.Get(`fields.#(name="y")`)
-			require.Equal(t, "y", prop.Get("name").String())
-			require.Equal(t, "", prop.Get("description").String())
-		})
-	}
-}
-
-func (ModuleSuite) TestGoWeirdFields(ctx context.Context, t *testctx.T) {
-	// these are all cases that used to panic due to the disparity in the type spec and the ast
-
-	c := connect(ctx, t)
-
-	modGen := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-type Z string
-
-type Minimal struct {
-	// field with single (normal) name
-	W string
-
-	// field with multiple names
-	X, Y string
-
-	// field with no names
-	Z
-}
-
-func New() Minimal {
-	return Minimal{
-		W: "-",
-		X: "-",
-		Y: "-",
-		Z: Z("-"),
-	}
-}
-
-// struct with no fields
-type Bar struct{}
-
-func (m *Minimal) Say(
-	// field with single (normal) name
-	a string,
-	// field with multiple names
-	b, c string,
-	// field with no names (not included, mixed names not allowed)
-	// string
-) string {
-	return a + " " + b + " " + c
-}
-
-func (m *Minimal) Hello(
-	// field with no names
-	string,
-) string {
-	return "hello"
-}
-
-func (m *Minimal) SayOpts(opts struct{
-	// field with single (normal) name
-	A string
-	// field with multiple names
-	B, C string
-	// field with no names (not included because of above)
-	// string
-}) string {
-	return opts.A + " " + opts.B + " " + opts.C
-}
-
-func (m *Minimal) HelloOpts(opts struct{
-	// field with no names
-	string
-}) string {
-	return "hello"
-}
-`,
-		)
-
-	out, err := modGen.With(daggerQuery(`{minimal{w, x, y, z}}`)).Stdout(ctx)
-	require.NoError(t, err)
-	require.JSONEq(t, `{"minimal": {"w": "-", "x": "-", "y": "-", "z": "-"}}`, out)
-
-	for _, name := range []string{"say", "sayOpts"} {
-		out, err := modGen.With(daggerQuery(`{minimal{%s(a: "hello", b: "world", c: "!")}}`, name)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, fmt.Sprintf(`{"minimal": {"%s": "hello world !"}}`, name), out)
-	}
-
-	for _, name := range []string{"hello", "helloOpts"} {
-		out, err := modGen.With(daggerQuery(`{minimal{%s(string: "")}}`, name)).Stdout(ctx)
-		require.NoError(t, err)
-		require.JSONEq(t, fmt.Sprintf(`{"minimal": {"%s": "hello"}}`, name), out)
-	}
-}
-
-func (ModuleSuite) TestGoFieldMustBeNil(ctx context.Context, t *testctx.T) {
-	c := connect(ctx, t)
-
-	modGen := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-import (
-	"fmt"
-	"dagger/minimal/internal/dagger"
-)
-
-type Minimal struct {
-	Src *dagger.Directory
-	Name *string
-}
-
-func New() *Minimal {
-	return &Minimal{}
-}
-
-func (m *Minimal) IsEmpty() bool {
-	if m.Name != nil {
-		panic(fmt.Sprintf("name should be nil but is %v", m.Name))
-	}
-	if m.Src != nil {
-		panic(fmt.Sprintf("src should be nil but is %v", m.Src))
-	}
-	return true
-}
-`,
-		)
-
-	out, err := modGen.With(daggerQuery(`{minimal{isEmpty}}`)).Stdout(ctx)
-	require.NoError(t, err)
-	require.JSONEq(t, `{"minimal": {"isEmpty": true}}`, out)
-}
-
-func (ModuleSuite) TestGoJSONField(ctx context.Context, t *testctx.T) {
-	c := connect(ctx, t)
-
-	modGen := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-import (
-	"dagger/minimal/internal/dagger"
-)
-
-type Minimal struct {
-	Config dagger.JSON
-}
-
-func New() *Minimal {
-	return &Minimal{
-		Config: "{\"a\":1}",
-	}
-}
-`,
-		)
-
-	out, err := modGen.With(daggerQuery(`{minimal{config}}`)).Stdout(ctx)
-	require.NoError(t, err)
-	require.JSONEq(t, `{"minimal":{"config":"{\"a\":1}"}}`, out)
 }
 
 func (ModuleSuite) TestDescription(ctx context.Context, t *testctx.T) {
@@ -2113,701 +530,6 @@ class Test {
 	}
 }
 
-// this is no longer allowed, but verify the SDK errors out
-func (ModuleSuite) TestGoExtendCore(ctx context.Context, t *testctx.T) {
-	moreContents := `package dagger
-
-import (
-	"context"
-)
-
-func (c *Container) Echo(ctx context.Context, msg string) (string, error) {
-	return c.WithExec([]string{"echo", msg}).Stdout(ctx)
-}
-`
-
-	t.Run("in different mod name", func(ctx context.Context, t *testctx.T) {
-		var logs safeBuffer
-		c := connect(ctx, t, dagger.WithLogOutput(&logs))
-		_, err := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--source=.", "--name=test", "--sdk=go")).
-			WithNewFile("/work/internal/dagger/more.go", moreContents).
-			With(daggerQuery(`{container{from(address:"` + alpineImage + `"){echo(msg:"echo!"){stdout}}}}`)).
-			Sync(ctx)
-		require.Error(t, err)
-		require.NoError(t, c.Close())
-		t.Log(logs.String())
-		require.Regexp(t, "cannot define methods on objects from outside this module", logs.String())
-	})
-
-	t.Run("in same mod name", func(ctx context.Context, t *testctx.T) {
-		var logs safeBuffer
-		c := connect(ctx, t, dagger.WithLogOutput(&logs))
-		_, err := c.Container().From(golangImage).
-			WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-			WithWorkdir("/work").
-			With(daggerExec("init", "--source=.", "--name=container", "--sdk=go")).
-			WithNewFile("/work/internal/dagger/more.go", moreContents).
-			With(daggerQuery(`{container{from(address:"` + alpineImage + `"){echo(msg:"echo!"){stdout}}}}`)).
-			Sync(ctx)
-		require.Error(t, err)
-		require.NoError(t, c.Close())
-		t.Log(logs.String())
-		require.Regexp(t, "cannot define methods on objects from outside this module", logs.String())
-	})
-}
-
-func (ModuleSuite) TestGoBadCtx(ctx context.Context, t *testctx.T) {
-	var logs safeBuffer
-	c := connect(ctx, t, dagger.WithLogOutput(&logs))
-
-	_, err := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=foo", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-import "context"
-
-type Foo struct {}
-
-func (f *Foo) Echo(ctx context.Context, ctx2 context.Context) (string, error) {
-	return "", nil
-}
-`,
-		).
-		With(daggerQuery(`{foo{echo}}`)).
-		Sync(ctx)
-	require.Error(t, err)
-	require.NoError(t, c.Close())
-	t.Log(logs.String())
-	require.Regexp(t, "unexpected context type", logs.String())
-}
-
-func (ModuleSuite) TestCustomTypes(ctx context.Context, t *testctx.T) {
-	type testCase struct {
-		sdk    string
-		source string
-	}
-
-	for _, tc := range []testCase{
-		{
-			sdk: "go",
-			source: `package main
-
-import "strings"
-
-type Test struct{}
-
-func (m *Test) Repeater(msg string, times int) *Repeater {
-	return &Repeater{
-		Message: msg,
-		Times:   times,
-	}
-}
-
-type Repeater struct {
-	Message string
-	Times   int
-}
-
-func (t *Repeater) Render() string {
-	return strings.Repeat(t.Message, t.Times)
-}
-`,
-		},
-		{
-			sdk: "python",
-			source: `from dagger import field, function, object_type
-
-@object_type
-class Repeater:
-    message: str = field(default="")
-    times: int = field(default=0)
-
-    @function
-    def render(self) -> str:
-        return self.message * self.times
-
-@function
-def repeater(msg: str, times: int) -> Repeater:
-    return Repeater(message=msg, times=times)
-`,
-		},
-		{
-			sdk: "typescript",
-			source: `
-import { object, func } from "@dagger.io/dagger"
-
-@object()
-class Repeater {
-  @func()
-  message: string
-
-  @func()
-  times: number
-
-  constructor(message: string, times: number) {
-    this.message = message
-    this.times = times
-  }
-
-  @func()
-  render(): string {
-    return this.message.repeat(this.times)
-  }
-}
-
-@object()
-class Test {
-  @func()
-  repeater(msg: string, times: number): Repeater {
-    return new Repeater(msg, times)
-  }
-}
-`,
-		},
-	} {
-		tc := tc
-
-		t.Run(fmt.Sprintf("custom %s types", tc.sdk), func(ctx context.Context, t *testctx.T) {
-			c := connect(ctx, t)
-
-			out, err := modInit(t, c, tc.sdk, tc.source).
-				With(daggerQuery(`{test{repeater(msg:"echo!", times: 3){render}}}`)).
-				Stdout(ctx)
-
-			require.NoError(t, err)
-			require.JSONEq(t, `{"test":{"repeater":{"render":"echo!echo!echo!"}}}`, out)
-		})
-	}
-}
-
-func (ModuleSuite) TestReturnTypeDetection(ctx context.Context, t *testctx.T) {
-	type testCase struct {
-		sdk    string
-		source string
-	}
-
-	for _, tc := range []testCase{
-		{
-			sdk: "go",
-			source: `package main
-
-type Foo struct {}
-
-type X struct {
-	Message string ` + "`json:\"message\"`" + `
-}
-
-func (m *Foo) MyFunction() X {
-	return X{Message: "foo"}
-}
-`,
-		},
-		{
-			sdk: "python",
-			source: `from dagger import field, function, object_type
-
-@object_type
-class X:
-    message: str = field(default="")
-
-@function
-def my_function() -> X:
-    return X(message="foo")
-`,
-		},
-		{
-			sdk: "typescript",
-			source: `
-import { object, func } from "@dagger.io/dagger"
-
-@object()
-class X {
-  @func()
-  message: string
-
-  constructor(message: string) {
-    this.message = message;
-  }
-}
-
-@object()
-class Foo {
-  @func()
-  myFunction(): X {
-    return new X("foo");
-  }
-}
-`,
-		},
-	} {
-		tc := tc
-
-		t.Run(tc.sdk, func(ctx context.Context, t *testctx.T) {
-			c := connect(ctx, t)
-
-			modGen := c.Container().From(golangImage).
-				WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-				WithWorkdir("/work").
-				With(daggerExec("init", "--name=foo", "--sdk="+tc.sdk)).
-				With(sdkSource(tc.sdk, tc.source))
-
-			out, err := modGen.With(daggerQuery(`{foo{myFunction{message}}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.JSONEq(t, `{"foo":{"myFunction":{"message":"foo"}}}`, out)
-		})
-	}
-}
-
-func (ModuleSuite) TestReturnObject(ctx context.Context, t *testctx.T) {
-	type testCase struct {
-		sdk    string
-		source string
-	}
-
-	for _, tc := range []testCase{
-		{
-			sdk: "go",
-			source: `package main
-
-type Foo struct {}
-
-type X struct {
-	Message string ` + "`json:\"message\"`" + `
-	When string ` + "`json:\"Timestamp\"`" + `
-	To string ` + "`json:\"recipient\"`" + `
-	From string
-}
-
-func (m *Foo) MyFunction() X {
-	return X{Message: "foo", When: "now", To: "user", From: "admin"}
-}
-`,
-		},
-		{
-			sdk: "python",
-			source: `from dagger import field, function, object_type
-
-@object_type
-class X:
-    message: str = field(default="")
-    when: str = field(default="", name="Timestamp")
-    to: str = field(default="", name="recipient")
-    from_: str = field(default="", name="from")
-
-@object_type
-class Foo:
-    @function
-    def my_function(self) -> X:
-        return X(message="foo", when="now", to="user", from_="admin")
-`,
-		},
-		{
-			sdk: "typescript",
-			source: `
-import { object, func } from "@dagger.io/dagger"
-
-@object()
-class X {
-  @func()
-  message: string
-
-  @func()
-  timestamp: string
-
-  @func()
-  recipient: string
-
-  @func()
-  from: string
-
-  constructor(message: string, timestamp: string, recipient: string, from: string) {
-    this.message = message;
-    this.timestamp = timestamp;
-    this.recipient = recipient;
-    this.from = from;
-  }
-}
-
-@object()
-class Foo {
-  @func()
-  myFunction(): X {
-    return new X("foo", "now", "user", "admin");
-  }
-}
-`,
-		},
-	} {
-		tc := tc
-
-		t.Run(tc.sdk, func(ctx context.Context, t *testctx.T) {
-			c := connect(ctx, t)
-
-			modGen := c.Container().From(golangImage).
-				WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-				WithWorkdir("/work").
-				With(daggerExec("init", "--name=foo", "--sdk="+tc.sdk)).
-				With(sdkSource(tc.sdk, tc.source))
-
-			out, err := modGen.With(daggerQuery(`{foo{myFunction{message, recipient, from, timestamp}}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.JSONEq(t, `{"foo":{"myFunction":{"message":"foo", "recipient":"user", "from":"admin", "timestamp":"now"}}}`, out)
-		})
-	}
-}
-
-func (ModuleSuite) TestReturnNestedObject(ctx context.Context, t *testctx.T) {
-	type testCase struct {
-		sdk    string
-		source string
-	}
-
-	for _, tc := range []testCase{
-		{
-			sdk: "go",
-			source: `package main
-
-type Playground struct{}
-
-type Foo struct {
-	MsgContainer Bar
-}
-
-type Bar struct {
-	Msg string
-}
-
-func (m *Playground) MyFunction() Foo {
-	return Foo{MsgContainer: Bar{Msg: "hello world"}}
-}
-`,
-		},
-		{
-			sdk: "python",
-			source: `from dagger import field, function, object_type
-
-@object_type
-class Bar:
-    msg: str = field()
-
-@object_type
-class Foo:
-    msg_container: Bar = field()
-
-@object_type
-class Playground:
-    @function
-    def my_function(self) -> Foo:
-        return Foo(msg_container=Bar(msg="hello world"))
-`,
-		},
-		{
-			sdk: "typescript",
-			source: `
-import { object, func } from "@dagger.io/dagger"
-
-@object()
-class Bar {
-  @func()
-  msg: string;
-
-  constructor(msg: string) {
-    this.msg = msg;
-  }
-}
-
-@object()
-class Foo {
-  @func()
-  msgContainer: Bar;
-
-  constructor(msgContainer: Bar) {
-    this.msgContainer = msgContainer;
-  }
-}
-
-@object()
-class Playground {
-  @func()
-  myFunction(): Foo {
-    return new Foo(new Bar("hello world"));
-  }
-}
-`,
-		},
-	} {
-		tc := tc
-
-		t.Run(tc.sdk, func(ctx context.Context, t *testctx.T) {
-			c := connect(ctx, t)
-
-			modGen := c.Container().From(golangImage).
-				WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-				WithWorkdir("/work").
-				With(daggerExec("init", "--name=playground", "--sdk="+tc.sdk)).
-				With(sdkSource(tc.sdk, tc.source))
-
-			out, err := modGen.With(daggerQuery(`{playground{myFunction{msgContainer{msg}}}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.JSONEq(t, `{"playground":{"myFunction":{"msgContainer":{"msg": "hello world"}}}}`, out)
-		})
-	}
-}
-
-func (ModuleSuite) TestReturnCompositeCore(ctx context.Context, t *testctx.T) {
-	type testCase struct {
-		sdk    string
-		source string
-	}
-
-	for _, tc := range []testCase{
-		{
-			sdk: "go",
-			source: `package main
-
-import (
-	"dagger/playground/internal/dagger"
-)
-
-type Playground struct{}
-
-func (m *Playground) MySlice() []*dagger.Container {
-	return []*dagger.Container{dag.Container().From("` + alpineImage + `").WithExec([]string{"echo", "hello world"})}
-}
-
-type Foo struct {
-	Con *dagger.Container
-	// verify fields can remain nil w/out error too
-	UnsetFile *dagger.File
-}
-
-func (m *Playground) MyStruct() *Foo {
-	return &Foo{Con: dag.Container().From("` + alpineImage + `").WithExec([]string{"echo", "hello world"})}
-}
-`,
-		},
-		{
-			sdk: "python",
-			source: `import dagger
-from dagger import dag, field, function, object_type
-
-@object_type
-class Foo:
-    con: dagger.Container = field()
-    unset_file: dagger.File | None = field(default=None)
-
-@object_type
-class Playground:
-    @function
-    def my_slice(self) -> list[dagger.Container]:
-        return [dag.container().from_("` + alpineImage + `").with_exec(["echo", "hello world"])]
-
-    @function
-    def my_struct(self) -> Foo:
-        return Foo(con=dag.container().from_("` + alpineImage + `").with_exec(["echo", "hello world"]))
-`,
-		},
-		{
-			sdk: "typescript",
-			source: `
-import { dag, Container, File, object, func } from "@dagger.io/dagger"
-
-@object()
-class Foo {
-  @func()
-  con: Container
-
-  @func()
-  unsetFile?: File
-
-  constructor(con: Container, unsetFile?: File) {
-    this.con = con
-    this.unsetFile = unsetFile
-  }
-}
-
-@object()
-class Playground {
-  @func()
-  mySlice(): Container[] {
-    return [
-      dag.container().from("` + alpineImage + `").withExec(["echo", "hello world"])
-    ]
-  }
-
-  @func()
-  myStruct(): Foo {
-    return new Foo(
-      dag.container().from("` + alpineImage + `").withExec(["echo", "hello world"])
-    )
-  }
-}
-`,
-		},
-	} {
-		tc := tc
-
-		t.Run(tc.sdk, func(ctx context.Context, t *testctx.T) {
-			c := connect(ctx, t)
-
-			modGen := c.Container().From(golangImage).
-				WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-				WithWorkdir("/work").
-				With(daggerExec("init", "--name=playground", "--sdk="+tc.sdk)).
-				With(sdkSource(tc.sdk, tc.source))
-
-			out, err := modGen.With(daggerQuery(`{playground{mySlice{stdout}}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.JSONEq(t, `{"playground":{"mySlice":[{"stdout":"hello world\n"}]}}`, out)
-
-			out, err = modGen.With(daggerQuery(`{playground{myStruct{con{stdout}}}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.JSONEq(t, `{"playground":{"myStruct":{"con":{"stdout":"hello world\n"}}}}`, out)
-		})
-	}
-}
-
-func (ModuleSuite) TestReturnComplexThing(ctx context.Context, t *testctx.T) {
-	type testCase struct {
-		sdk    string
-		source string
-	}
-
-	for _, tc := range []testCase{
-		{
-			sdk: "go",
-			source: `package main
-
-import (
-	"dagger/playground/internal/dagger"
-)
-
-type Playground struct{}
-
-type ScanResult struct {
-	Containers	[]*dagger.Container ` + "`json:\"targets\"`" + `
-	Report		ScanReport
-}
-
-type ScanReport struct {
-	Contents string ` + "`json:\"contents\"`" + `
-	Authors  []string ` + "`json:\"Authors\"`" + `
-}
-
-func (m *Playground) Scan() ScanResult {
-	return ScanResult{
-		Containers: []*dagger.Container{
-			dag.Container().From("` + alpineImage + `").WithExec([]string{"echo", "hello world"}),
-		},
-		Report: ScanReport{
-			Contents: "hello world",
-			Authors: []string{"foo", "bar"},
-		},
-	}
-}
-`,
-		},
-		{
-			sdk: "python",
-			source: `import dagger
-from dagger import dag, field, function, object_type
-
-@object_type
-class ScanReport:
-    contents: str = field()
-    authors: list[str] = field()
-
-@object_type
-class ScanResult:
-    containers: list[dagger.Container] = field(name="targets")
-    report: ScanReport = field()
-
-@object_type
-class Playground:
-    @function
-    def scan(self) -> ScanResult:
-        return ScanResult(
-            containers=[
-                dag.container().from_("` + alpineImage + `").with_exec(["echo", "hello world"]),
-            ],
-            report=ScanReport(
-                contents="hello world",
-                authors=["foo", "bar"],
-            ),
-        )
-`,
-		},
-		{
-			sdk: "typescript",
-			source: `
-import { dag, Container, object, func } from "@dagger.io/dagger"
-
-@object()
-class ScanReport {
-  @func()
-  contents: string
-
-  @func()
-  authors: string[]
-
-  constructor(contents: string, authors: string[]) {
-    this.contents = contents
-    this.authors = authors
-  }
-}
-
-@object()
-class ScanResult {
-  @func("targets")
-  containers: Container[]
-
-  @func()
-  report: ScanReport
-
-  constructor(containers: Container[], report: ScanReport) {
-    this.containers = containers
-    this.report = report
-  }
-}
-
-@object()
-class Playground {
-  @func()
-  async scan(): Promise<ScanResult> {
-    return new ScanResult(
-      [
-        dag.container().from("` + alpineImage + `").withExec(["echo", "hello world"])
-      ],
-      new ScanReport("hello world", ["foo", "bar"])
-    )
-  }
-}
-`,
-		},
-	} {
-		tc := tc
-
-		t.Run(tc.sdk, func(ctx context.Context, t *testctx.T) {
-			c := connect(ctx, t)
-
-			modGen := c.Container().From(golangImage).
-				WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-				WithWorkdir("/work").
-				With(daggerExec("init", "--name=playground", "--sdk="+tc.sdk)).
-				With(sdkSource(tc.sdk, tc.source))
-
-			out, err := modGen.With(daggerQuery(`{playground{scan{targets{stdout},report{contents,authors}}}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.JSONEq(t, `{"playground":{"scan":{"targets":[{"stdout":"hello world\n"}],"report":{"contents":"hello world","authors":["foo","bar"]}}}}`, out)
-		})
-	}
-}
-
 func (ModuleSuite) TestGlobalVarDAG(ctx context.Context, t *testctx.T) {
 	type testCase struct {
 		sdk    string
@@ -2876,750 +598,6 @@ class Foo {
 			require.JSONEq(t, `{"foo":{"fn":"foo\n"}}`, out)
 		})
 	}
-}
-
-func (ModuleSuite) TestIDableType(ctx context.Context, t *testctx.T) {
-	type testCase struct {
-		sdk    string
-		source string
-	}
-
-	for _, tc := range []testCase{
-		{
-			sdk: "go",
-			source: `package main
-
-type Foo struct {
-	Data string
-}
-
-func (m *Foo) Set(data string) *Foo {
-	m.Data = data
-	return m
-}
-
-func (m *Foo) Get() string {
-	return m.Data
-}
-`,
-		},
-		{
-			sdk: "python",
-			source: `from typing import Self
-
-from dagger import field, function, object_type
-
-@object_type
-class Foo:
-    data: str = ""
-
-    @function
-    def set(self, data: str) -> Self:
-        self.data = data
-        return self
-
-    @function
-    def get(self) -> str:
-        return self.data
-`,
-		},
-		{
-			sdk: "typescript",
-			source: `
-import { object, func } from "@dagger.io/dagger"
-
-@object()
-class Foo {
-  data: string = ""
-
-  @func()
-  set(data: string): Foo {
-    this.data = data
-    return this
-  }
-
-  @func()
-  get(): string {
-    return this.data
-  }
-}
-`,
-		},
-	} {
-		tc := tc
-
-		t.Run(tc.sdk, func(ctx context.Context, t *testctx.T) {
-			c := connect(ctx, t)
-
-			modGen := c.Container().From(golangImage).
-				WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-				WithWorkdir("/work").
-				With(daggerExec("init", "--name=foo", "--sdk="+tc.sdk)).
-				With(sdkSource(tc.sdk, tc.source))
-
-			// sanity check
-			out, err := modGen.With(daggerQuery(`{foo{set(data: "abc"){get}}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.JSONEq(t, `{"foo":{"set":{"get": "abc"}}}`, out)
-
-			out, err = modGen.With(daggerQuery(`{foo{set(data: "abc"){id}}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			id := gjson.Get(out, "foo.set.id").String()
-
-			var idp call.ID
-			err = idp.Decode(id)
-			require.NoError(t, err)
-			require.Equal(t, idp.Display(), `foo.set(data: "abc"): Foo!`)
-
-			out, err = modGen.With(daggerQuery(`{loadFooFromID(id: "%s"){get}}`, id)).Stdout(ctx)
-			require.NoError(t, err)
-			require.JSONEq(t, `{"loadFooFromID":{"get": "abc"}}`, out)
-		})
-	}
-}
-
-func (ModuleSuite) TestArgOwnType(ctx context.Context, t *testctx.T) {
-	// Verify use of a module's own object as an argument type.
-	// The server needs to specifically decode the input type from an ID into
-	// the raw JSON, since the module doesn't understand it's own types as IDs
-
-	type testCase struct {
-		sdk    string
-		source string
-	}
-	for _, tc := range []testCase{
-		{
-			sdk: "go",
-			source: `package main
-
-import "strings"
-
-type Foo struct{}
-
-type Message struct {
-	Content string
-}
-
-func (m *Foo) SayHello(name string) Message {
-	return Message{Content: "hello " + name}
-}
-
-func (m *Foo) Upper(msg Message) Message {
-	msg.Content = strings.ToUpper(msg.Content)
-	return msg
-}
-
-func (m *Foo) Uppers(msg []Message) []Message {
-	for i := range msg {
-		msg[i].Content = strings.ToUpper(msg[i].Content)
-	}
-	return msg
-}`,
-		},
-		{
-			sdk: "python",
-			source: `from dagger import field, function, object_type
-
-@object_type
-class Message:
-    content: str = field()
-
-@object_type
-class Foo:
-    @function
-    def say_hello(self, name: str) -> Message:
-        return Message(content=f"hello {name}")
-
-    @function
-    def upper(self, msg: Message) -> Message:
-        msg.content = msg.content.upper()
-        return msg
-
-    @function
-    def uppers(self, msg: list[Message]) -> list[Message]:
-        for m in msg:
-            m.content = m.content.upper()
-        return msg
-`,
-		},
-		{
-			sdk: "typescript",
-			source: `
-import { object, func } from "@dagger.io/dagger"
-
-@object()
-class Message {
-  @func()
-  content: string
-
-  constructor(content: string) {
-    this.content = content
-  }
-}
-
-@object()
-class Foo {
-  @func()
-  sayHello(name: string): Message {
-    return new Message("hello " + name)
-  }
-
-  @func()
-  upper(msg: Message): Message {
-    msg.content = msg.content.toUpperCase()
-    return msg
-  }
-
-  @func()
-  uppers(msg: Message[]): Message[] {
-    for (let i = 0; i < msg.length; i++) {
-      msg[i].content = msg[i].content.toUpperCase()
-    }
-    return msg
-  }
-}
-`,
-		},
-	} {
-		tc := tc
-
-		t.Run(tc.sdk, func(ctx context.Context, t *testctx.T) {
-			c := connect(ctx, t)
-
-			modGen := c.Container().From(golangImage).
-				WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-				WithWorkdir("/work").
-				With(daggerExec("init", "--name=foo", "--sdk="+tc.sdk)).
-				With(sdkSource(tc.sdk, tc.source))
-
-			out, err := modGen.With(daggerQuery(`{foo{sayHello(name: "world"){id}}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			id := gjson.Get(out, "foo.sayHello.id").String()
-			var idp call.ID
-			err = idp.Decode(id)
-			require.NoError(t, err)
-			require.Equal(t, idp.Display(), `foo.sayHello(name: "world"): FooMessage!`)
-
-			out, err = modGen.With(daggerQuery(`{foo{upper(msg:"%s"){content}}}`, id)).Stdout(ctx)
-			require.NoError(t, err)
-			require.JSONEq(t, `{"foo":{"upper":{"content": "HELLO WORLD"}}}`, out)
-
-			out, err = modGen.With(daggerQuery(`{foo{uppers(msg:["%s", "%s"]){content}}}`, id, id)).Stdout(ctx)
-			require.NoError(t, err)
-			require.JSONEq(t, `{"foo":{"uppers":[{"content": "HELLO WORLD"}, {"content": "HELLO WORLD"}]}}`, out)
-		})
-	}
-}
-
-func (ModuleSuite) TestScalarType(ctx context.Context, t *testctx.T) {
-	type testCase struct {
-		sdk    string
-		source string
-	}
-	for _, tc := range []testCase{
-		{
-			sdk: "go",
-			source: `package main
-
-import "dagger/test/internal/dagger"
-
-type Test struct{}
-
-func (m *Test) FromPlatform(platform dagger.Platform) string {
-	return string(platform)
-}
-
-func (m *Test) ToPlatform(platform string) dagger.Platform {
-	return dagger.Platform(platform)
-}
-
-func (m *Test) FromPlatforms(platform []dagger.Platform) []string {
-	result := []string{}
-	for _, p := range platform {
-		result = append(result, string(p))
-	}
-	return result
-}
-
-func (m *Test) ToPlatforms(platform []string) []dagger.Platform {
-	result := []dagger.Platform{}
-	for _, p := range platform {
-		result = append(result, dagger.Platform(p))
-	}
-	return result
-}
-`,
-		},
-		{
-			sdk: "python",
-			source: `import dagger
-from dagger import function, object_type
-
-@object_type
-class Test:
-    @function
-    def from_platform(self, platform: dagger.Platform) -> str:
-        return str(platform)
-
-    @function
-    def to_platform(self, platform: str) -> dagger.Platform:
-        return dagger.Platform(platform)
-
-    @function
-    def from_platforms(self, platform: list[dagger.Platform]) -> list[str]:
-        return [str(p) for p in platform]
-
-    @function
-    def to_platforms(self, platform: list[str]) -> list[dagger.Platform]:
-        return [dagger.Platform(p) for p in platform]
-`,
-		},
-		{
-			sdk: "typescript",
-			source: `import { object, func, Platform } from "@dagger.io/dagger"
-
-@object()
-class Test {
-	@func()
-	fromPlatform(platform: Platform): string {
-		return platform as string
-	}
-
-	@func()
-	toPlatform(platform: string): Platform {
-		return platform as Platform
-	}
-
-	@func()
-	fromPlatforms(platform: Platform[]): string[] {
-		return platform.map(p => p as string)
-	}
-
-	@func()
-	toPlatforms(platform: string[]): Platform[] {
-		return platform.map(p => p as Platform)
-	}
-}
-`,
-		},
-	} {
-		tc := tc
-
-		t.Run(tc.sdk, func(ctx context.Context, t *testctx.T) {
-			c := connect(ctx, t)
-			modGen := modInit(t, c, tc.sdk, tc.source)
-
-			out, err := modGen.With(daggerQuery(`{test{fromPlatform(platform: "linux/amd64")}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.Equal(t, "linux/amd64", gjson.Get(out, "test.fromPlatform").String())
-			_, err = modGen.With(daggerQuery(`{test{fromPlatform(platform: "invalid")}}`)).Stdout(ctx)
-			require.ErrorContains(t, err, "unknown operating system or architecture")
-
-			out, err = modGen.With(daggerQuery(`{test{toPlatform(platform: "linux/amd64")}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.Equal(t, "linux/amd64", gjson.Get(out, "test.toPlatform").String())
-			_, err = modGen.With(daggerQuery(`{test{toPlatform(platform: "invalid")}}`)).Sync(ctx)
-			require.ErrorContains(t, err, "unknown operating system or architecture")
-
-			out, err = modGen.With(daggerQuery(`{test{fromPlatforms(platform: ["linux/amd64"])}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.Equal(t, 1, len(gjson.Get(out, "test.fromPlatforms").Array()))
-			require.Equal(t, "linux/amd64", gjson.Get(out, "test.fromPlatforms.0").String())
-			_, err = modGen.With(daggerQuery(`{test{fromPlatforms(platform: ["invalid"])}}`)).Stdout(ctx)
-			require.ErrorContains(t, err, "unknown operating system or architecture")
-
-			out, err = modGen.With(daggerQuery(`{test{toPlatforms(platform: ["linux/amd64"])}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.Equal(t, 1, len(gjson.Get(out, "test.toPlatforms.0").Array()))
-			require.Equal(t, "linux/amd64", gjson.Get(out, "test.toPlatforms.0").String())
-			_, err = modGen.With(daggerQuery(`{test{toPlatforms(platform: ["invalid"])}}`)).Sync(ctx)
-			require.ErrorContains(t, err, "unknown operating system or architecture")
-		})
-	}
-}
-
-func (ModuleSuite) TestEnumType(ctx context.Context, t *testctx.T) {
-	type testCase struct {
-		sdk    string
-		source string
-	}
-	for _, tc := range []testCase{
-		{
-			sdk: "go",
-			source: `package main
-
-import "dagger/test/internal/dagger"
-
-type Test struct{}
-
-func (m *Test) FromProto(proto dagger.NetworkProtocol) string {
-	return string(proto)
-}
-
-func (m *Test) ToProto(proto string) dagger.NetworkProtocol {
-	return dagger.NetworkProtocol(proto)
-}
-`,
-		},
-		{
-			sdk: "python",
-			source: `import dagger
-from dagger import function, object_type
-
-@object_type
-class Test:
-    @function
-    def from_proto(self, proto: dagger.NetworkProtocol) -> str:
-        return str(proto)
-
-    @function
-    def to_proto(self, proto: str) -> dagger.NetworkProtocol:
-        # Doing "dagger.NetworkProtocol(proto)" will fail in Python, so mock
-        # it to force sending the invalid value back to the server.
-        from dagger.client.base import Enum
-
-        class MockEnum(Enum):
-            TCP = "TCP"
-            INVALID = "INVALID"
-
-        return MockEnum(proto)
-`,
-		},
-		{
-			sdk: "typescript",
-			source: `import { object, func, NetworkProtocol } from "@dagger.io/dagger";
-
-@object()
-class Test {
-  @func()
-  fromProto(Proto: NetworkProtocol): string {
-    return Proto as string;
-  }
-
-  @func()
-  toProto(Proto: string): NetworkProtocol {
-    return Proto as NetworkProtocol;
-  }
-}
-`,
-		},
-	} {
-		tc := tc
-
-		t.Run(tc.sdk, func(ctx context.Context, t *testctx.T) {
-			c := connect(ctx, t)
-			modGen := modInit(t, c, tc.sdk, tc.source)
-
-			out, err := modGen.With(daggerQuery(`{test{fromProto(proto: "TCP")}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.Equal(t, "TCP", gjson.Get(out, "test.fromProto").String())
-
-			_, err = modGen.With(daggerQuery(`{test{fromProto(proto: "INVALID")}}`)).Stdout(ctx)
-			require.ErrorContains(t, err, "invalid enum value")
-
-			out, err = modGen.With(daggerQuery(`{test{toProto(proto: "TCP")}}`)).Stdout(ctx)
-			require.NoError(t, err)
-			require.Equal(t, "TCP", gjson.Get(out, "test.toProto").String())
-
-			_, err = modGen.With(daggerQuery(`{test{toProto(proto: "INVALID")}}`)).Sync(ctx)
-			require.ErrorContains(t, err, "invalid enum value")
-		})
-	}
-}
-
-func (ModuleSuite) TestCustomEnumType(ctx context.Context, t *testctx.T) {
-	t.Run("custom enum type", func(ctx context.Context, t *testctx.T) {
-		type testCase struct {
-			sdk    string
-			source string
-		}
-		for _, tc := range []testCase{
-			{
-				sdk: "go",
-				source: `package main
-
-// Enum for Status
-type Status string
-
-const (
-	// Active status
-	Active Status = "ACTIVE"
-
-	// Inactive status
-	Inactive Status = "INACTIVE"
-)
-
-func New(
-	// +default="INACTIVE"
-	status Status,
-) *Test {
-	return &Test{Status: status}
-}
-
-type Test struct {
-	Status Status
-}
-
-func (m *Test) FromStatus(status Status) string {
-	return string(status)
-}
-
-func (m *Test) FromStatusOpt(
-	// +optional
-	status Status,
-) string {
-	return string(status)
-}
-
-func (m *Test) ToStatus(status string) Status {
-	return Status(status)
-}
-`,
-			},
-			{
-				sdk: "python",
-				source: `import dagger
-
-@dagger.enum_type
-class Status(dagger.Enum):
-    """Enum for Status"""
-
-    ACTIVE = "ACTIVE", "Active status"
-    INACTIVE = "INACTIVE", "Inactive status"
-
-
-@dagger.object_type
-class Test:
-    status: Status = dagger.field(default=Status.INACTIVE)
-
-    @dagger.function
-    def from_status(self, status: Status) -> str:
-        return str(status)
-
-    @dagger.function
-    def from_status_opt(self, status: Status | None) -> str:
-        return str(status) if status else ""
-
-    @dagger.function
-    def to_status(self, status: str) -> Status:
-        # Doing "Status(proto)" will fail in Python, so mock
-        # it to force sending the invalid value back to the server.
-        class MockEnum(dagger.Enum):
-            INACTIVE = "INACTIVE"
-            INVALID = "INVALID"
-
-        return MockEnum(status)
-`,
-			},
-			{
-				sdk: "typescript",
-				source: `import { func, object, enumType } from "@dagger.io/dagger"
-
-/**
- * Enum for Status
- */
-@enumType()
-class Status {
-  /**
-   * Active status
-   */
-  static readonly Active: string = "ACTIVE"
-
-  /**
-   * Inactive status
-   */
-  static readonly Inactive: string = "INACTIVE"
-}
-
-@object()
-export class Test {
-  @func()
-  status: Status
-
-  // FIXME: this should be Status.Inactive instead of "INACTIVE"
-  constructor(status: Status = "INACTIVE") {
-    this.status = status
-  }
-
-  @func()
-  fromStatus(status: Status): string {
-    return status as string
-  }
-
-  @func()
-  fromStatusOpt(status?: Status): string {
-	if (status) {
-		return status as string
-	}
-    return ""
-  }
-
-  @func()
-  toStatus(status: string): Status {
-    return status as Status
-  }
-}
-`,
-			},
-		} {
-			tc := tc
-
-			t.Run(tc.sdk, func(ctx context.Context, t *testctx.T) {
-				c := connect(ctx, t)
-				modGen := modInit(t, c, tc.sdk, tc.source)
-
-				// fromStatus
-				out, err := modGen.With(daggerQuery(`{test{fromStatus(status: "ACTIVE")}}`)).Stdout(ctx)
-				require.NoError(t, err)
-				require.Equal(t, "ACTIVE", gjson.Get(out, "test.fromStatus").String())
-
-				out, err = modGen.With(daggerQuery(`{test{status}}`)).Stdout(ctx)
-				require.NoError(t, err)
-				require.Equal(t, "INACTIVE", gjson.Get(out, "test.status").String())
-
-				_, err = modGen.With(daggerQuery(`{test{fromStatus(status: "INVALID")}}`)).Stdout(ctx)
-				require.ErrorContains(t, err, "invalid enum value")
-
-				// fromStatusOpt
-				out, err = modGen.With(daggerQuery(`{test{fromStatusOpt}}`)).Stdout(ctx)
-				require.NoError(t, err)
-				require.Equal(t, "", gjson.Get(out, "test.fromStatusOpt").String())
-
-				out, err = modGen.With(daggerQuery(`{test{fromStatusOpt(status: "ACTIVE")}}`)).Stdout(ctx)
-				require.NoError(t, err)
-				require.Equal(t, "ACTIVE", gjson.Get(out, "test.fromStatusOpt").String())
-
-				_, err = modGen.With(daggerQuery(`{test{fromStatusOpt(status: "INVALID")}}`)).Stdout(ctx)
-				require.ErrorContains(t, err, "invalid enum value")
-
-				// toStatus
-				out, err = modGen.With(daggerQuery(`{test{toStatus(status: "INACTIVE")}}`)).Stdout(ctx)
-				require.NoError(t, err)
-				require.Equal(t, "INACTIVE", gjson.Get(out, "test.toStatus").String())
-
-				_, err = modGen.With(daggerQuery(`{test{toStatus(status: "INVALID")}}`)).Sync(ctx)
-				require.ErrorContains(t, err, "invalid enum value")
-
-				// introspection
-				mod := inspectModule(ctx, t, modGen)
-				statusEnum := mod.Get("enums.#.asEnum|#(name=TestStatus)")
-				require.Equal(t, "Enum for Status", statusEnum.Get("description").String())
-				require.Len(t, statusEnum.Get("values").Array(), 2)
-				require.Equal(t, "ACTIVE", statusEnum.Get("values.0.name").String())
-				require.Equal(t, "INACTIVE", statusEnum.Get("values.1.name").String())
-				require.Equal(t, "Active status", statusEnum.Get("values.0.description").String())
-				require.Equal(t, "Inactive status", statusEnum.Get("values.1.description").String())
-			})
-		}
-	})
-
-	t.Run("custom external enum type", func(ctx context.Context, t *testctx.T) {
-		depSrc := `package main
-
-// Enum for Status
-type Status string
-
-const (
-	// Active status
-	Active Status = "ACTIVE"
-
-	// Inactive status
-	Inactive Status = "INACTIVE"
-)
-
-type Dep struct{}
-
-func (m *Dep) Active() Status {
-	return Active
-}
-
-func (m *Dep) Inactive() Status {
-	return Inactive
-}
-
-func (m *Dep) Invert(status Status) Status {
-	switch status {
-	case Active:
-		return Inactive
-	case Inactive:
-		return Active
-	default:
-		panic("invalid status")
-	}
-}
-`
-
-		type testCase struct {
-			sdk    string
-			source string
-		}
-		for _, tc := range []testCase{
-			{
-				sdk: "go",
-				source: `package main
-
-import "context"
-
-type Test struct{}
-
-func (m *Test) Test(ctx context.Context) (string, error) {
-	status, err := dag.Dep().Active(ctx)
-	if err != nil {
-		return "", err
-	}
-	status, err = dag.Dep().Invert(ctx, status)
-	if err != nil {
-		return "", err
-	}
-	return string(status), nil
-}
-`,
-			},
-			{
-				sdk: "python",
-				source: `import dagger
-from dagger import dag
-
-@dagger.object_type
-class Test:
-    @dagger.function
-    async def test(self) -> str:
-        status = await dag.dep().active()
-        status = await dag.dep().invert(status)
-        return str(status)
-`,
-			},
-			{
-				sdk: "typescript",
-				source: `import { dag, func, object } from "@dagger.io/dagger"
-
-@object()
-export class Test {
-  @func()
-  async test(): Promise<string> {
-    let status = await dag.dep().active();
-    status = await dag.dep().invert(status);
-    return status;
-  }
-}
-`,
-			},
-		} {
-			tc := tc
-
-			t.Run(tc.sdk, func(ctx context.Context, t *testctx.T) {
-				c := connect(ctx, t)
-
-				modGen := modInit(t, c, tc.sdk, tc.source).
-					With(withModInitAt("./dep", "go", depSrc)).
-					With(daggerExec("install", "./dep"))
-
-				out, err := modGen.With(daggerQuery(`{test{test}}`)).Stdout(ctx)
-				require.NoError(t, err)
-				require.Equal(t, "INACTIVE", gjson.Get(out, "test.test").String())
-			})
-		}
-	})
 }
 
 func (ModuleSuite) TestConflictingSameNameDeps(ctx context.Context, t *testctx.T) {
@@ -3779,297 +757,6 @@ func (m *Test) FnB() string {
 		Stdout(ctx)
 	require.NoError(t, err)
 	require.JSONEq(t, `{"test":{"fnA": "hi from b"}}`, out)
-}
-
-func (ModuleSuite) TestGoWithOtherModuleTypes(ctx context.Context, t *testctx.T) {
-	c := connect(ctx, t)
-
-	ctr := goGitBase(t, c).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work/dep").
-		With(daggerExec("init", "--source=.", "--name=dep", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-type Dep struct{}
-
-type Obj struct {
-	Foo string
-}
-
-func (m *Dep) Fn() Obj {
-	return Obj{Foo: "foo"}
-}
-`,
-		).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=test", "--name=test", "--sdk=go", "test")).
-		With(daggerExec("install", "-m=test", "./dep")).
-		WithWorkdir("/work/test")
-
-	t.Run("return as other module object", func(ctx context.Context, t *testctx.T) {
-		t.Run("direct", func(ctx context.Context, t *testctx.T) {
-			_, err := ctr.
-				WithNewFile("main.go", `package main
-
-import "dagger/test/internal/dagger"
-
-type Test struct{}
-
-func (m *Test) Fn() (*dagger.DepObj, error) {
-	return nil, nil
-}
-`,
-				).
-				With(daggerFunctions()).
-				Stdout(ctx)
-			require.Error(t, err)
-			require.ErrorContains(t, err, fmt.Sprintf(
-				"object %q function %q cannot return external type from dependency module %q",
-				"Test", "Fn", "dep",
-			))
-		})
-
-		t.Run("list", func(ctx context.Context, t *testctx.T) {
-			_, err := ctr.
-				WithNewFile("main.go", `package main
-
-import "dagger/test/internal/dagger"
-
-type Test struct{}
-
-func (m *Test) Fn() ([]*dagger.DepObj, error) {
-	return nil, nil
-}
-`,
-				).
-				With(daggerFunctions()).
-				Stdout(ctx)
-			require.Error(t, err)
-			require.ErrorContains(t, err, fmt.Sprintf(
-				"object %q function %q cannot return external type from dependency module %q",
-				"Test", "Fn", "dep",
-			))
-		})
-	})
-
-	t.Run("arg as other module object", func(ctx context.Context, t *testctx.T) {
-		t.Run("direct", func(ctx context.Context, t *testctx.T) {
-			_, err := ctr.WithNewFile("main.go", `package main
-
-import "dagger/test/internal/dagger"
-
-type Test struct{}
-
-func (m *Test) Fn(obj *dagger.DepObj) error {
-	return nil
-}
-`,
-			).
-				With(daggerFunctions()).
-				Stdout(ctx)
-			require.Error(t, err)
-			require.ErrorContains(t, err, fmt.Sprintf(
-				"object %q function %q arg %q cannot reference external type from dependency module %q",
-				"Test", "Fn", "obj", "dep",
-			))
-		})
-
-		t.Run("list", func(ctx context.Context, t *testctx.T) {
-			_, err := ctr.WithNewFile("main.go", `package main
-
-import "dagger/test/internal/dagger"
-
-type Test struct{}
-
-func (m *Test) Fn(obj []*dagger.DepObj) error {
-	return nil
-}
-`,
-			).
-				With(daggerFunctions()).
-				Stdout(ctx)
-			require.Error(t, err)
-			require.ErrorContains(t, err, fmt.Sprintf(
-				"object %q function %q arg %q cannot reference external type from dependency module %q",
-				"Test", "Fn", "obj", "dep",
-			))
-		})
-	})
-
-	t.Run("field as other module object", func(ctx context.Context, t *testctx.T) {
-		t.Run("direct", func(ctx context.Context, t *testctx.T) {
-			_, err := ctr.
-				WithNewFile("main.go", `package main
-
-import "dagger/test/internal/dagger"
-
-type Test struct{}
-
-type Obj struct {
-	Foo *dagger.DepObj
-}
-
-func (m *Test) Fn() (*Obj, error) {
-	return nil, nil
-}
-`,
-				).
-				With(daggerFunctions()).
-				Stdout(ctx)
-			require.Error(t, err)
-			require.ErrorContains(t, err, fmt.Sprintf(
-				"object %q field %q cannot reference external type from dependency module %q",
-				"Obj", "Foo", "dep",
-			))
-		})
-
-		t.Run("list", func(ctx context.Context, t *testctx.T) {
-			_, err := ctr.
-				WithNewFile("main.go", `package main
-
-import "dagger/test/internal/dagger"
-
-type Test struct{}
-
-type Obj struct {
-	Foo []*dagger.DepObj
-}
-
-func (m *Test) Fn() (*Obj, error) {
-	return nil, nil
-}
-`,
-				).
-				With(daggerFunctions()).
-				Stdout(ctx)
-			require.Error(t, err)
-			require.ErrorContains(t, err, fmt.Sprintf(
-				"object %q field %q cannot reference external type from dependency module %q",
-				"Obj", "Foo", "dep",
-			))
-		})
-	})
-}
-
-func (ModuleSuite) TestGoUseDaggerTypesDirect(ctx context.Context, t *testctx.T) {
-	var logs safeBuffer
-	c := connect(ctx, t, dagger.WithLogOutput(&logs))
-
-	modGen := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-import "dagger/minimal/internal/dagger"
-
-type Minimal struct{}
-
-func (m *Minimal) Foo(dir *dagger.Directory) (*dagger.Directory) {
-	return dir.WithNewFile("foo", "xxx")
-}
-
-func (m *Minimal) Bar(dir *dagger.Directory) (*dagger.Directory) {
-	return dir.WithNewFile("bar", "yyy")
-}
-
-`,
-		)
-
-	out, err := modGen.With(daggerQuery(`{directory{id}}`)).Stdout(ctx)
-	require.NoError(t, err)
-	dirID := gjson.Get(out, "directory.id").String()
-
-	out, err = modGen.With(daggerQuery(`{minimal{foo(dir: "%s"){file(path: "foo"){contents}}}}`, dirID)).Stdout(ctx)
-	require.NoError(t, err)
-	require.JSONEq(t, `{"minimal":{"foo":{"file":{"contents": "xxx"}}}}`, out)
-
-	out, err = modGen.With(daggerQuery(`{minimal{bar(dir: "%s"){file(path: "bar"){contents}}}}`, dirID)).Stdout(ctx)
-	require.NoError(t, err)
-	require.JSONEq(t, `{"minimal":{"bar":{"file":{"contents": "yyy"}}}}`, out)
-}
-
-func (ModuleSuite) TestGoUtilsPkg(ctx context.Context, t *testctx.T) {
-	var logs safeBuffer
-	c := connect(ctx, t, dagger.WithLogOutput(&logs))
-
-	modGen := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c)).
-		WithWorkdir("/work").
-		With(daggerExec("init", "--source=.", "--name=minimal", "--sdk=go")).
-		WithNewFile("main.go", `package main
-
-import (
-	"context"
-	"dagger/minimal/utils"
-)
-
-type Minimal struct{}
-
-func (m *Minimal) Hello(ctx context.Context) (string, error) {
-	return utils.Foo().File("foo").Contents(ctx)
-}
-
-`,
-		).
-		WithNewFile("utils/util.go", `package utils
-
-import "dagger/minimal/internal/dagger"
-
-func Foo() *dagger.Directory {
-	return dagger.Connect().Directory().WithNewFile("/foo", "hello world")
-}
-
-`,
-		)
-
-	out, err := modGen.With(daggerQuery(`{minimal{hello}}`)).Stdout(ctx)
-	require.NoError(t, err)
-	require.JSONEq(t, `{"minimal":{"hello":"hello world"}}`, out)
-}
-
-func (ModuleSuite) TestGoNameCase(ctx context.Context, t *testctx.T) {
-	c := connect(ctx, t)
-
-	ctr := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c))
-
-	ctr = ctr.
-		WithWorkdir("/toplevel/ssh").
-		With(daggerExec("init", "--name=ssh", "--sdk=go", "--source=.")).
-		WithNewFile("main.go", `package main
-
-type Ssh struct {}
-
-func (ssh *Ssh) SayHello() string {
-        return "hello!"
-}
-`,
-		)
-	out, err := ctr.With(daggerQuery(`{ssh{sayHello}}`)).Stdout(ctx)
-	require.NoError(t, err)
-	require.JSONEq(t, `{"ssh":{"sayHello":"hello!"}}`, out)
-
-	ctr = ctr.
-		WithWorkdir("/toplevel").
-		With(daggerExec("init", "--name=toplevel", "--sdk=go", "--source=.")).
-		With(daggerExec("install", "./ssh")).
-		WithNewFile("main.go", `package main
-
-import "context"
-
-type Toplevel struct {}
-
-func (t *Toplevel) SayHello(ctx context.Context) (string, error) {
-        return dag.SSH().SayHello(ctx)
-}
-`,
-		)
-	logGen(ctx, t, ctr.Directory("."))
-
-	out, err = ctr.With(daggerQuery(`{toplevel{sayHello}}`)).Stdout(ctx)
-	require.NoError(t, err)
-	require.JSONEq(t, `{"toplevel":{"sayHello":"hello!"}}`, out)
 }
 
 var useInner = `package main
@@ -4795,40 +1482,6 @@ class Test {
 		_, err = ctr.With(daggerCall("bar")).Sync(ctx)
 		require.NoError(t, err)
 	})
-}
-
-func (ModuleSuite) TestGoEmbedded(ctx context.Context, t *testctx.T) {
-	c := connect(ctx, t)
-
-	ctr := c.Container().From(golangImage).
-		WithMountedFile(testCLIBinPath, daggerCliFile(t, c))
-
-	ctr = ctr.
-		WithWorkdir("/playground").
-		With(daggerExec("init", "--name=playground", "--sdk=go", "--source=.")).
-		WithNewFile("main.go", `package main
-
-import (
-	"dagger/playground/internal/dagger"
-)
-
-type Playground struct {
-	*dagger.Directory
-}
-
-func New() Playground {
-	return Playground{Directory: dag.Directory()}
-}
-
-func (p *Playground) SayHello() string {
-	return "hello!"
-}
-`,
-		)
-
-	out, err := ctr.With(daggerQuery(`{playground{sayHello, directory{entries}}}`)).Stdout(ctx)
-	require.NoError(t, err)
-	require.JSONEq(t, `{"playground":{"sayHello":"hello!", "directory":{"entries": []}}}`, out)
 }
 
 func (ModuleSuite) TestWrapping(ctx context.Context, t *testctx.T) {
